==> 05e-gcb_healthy_pca_deseq.R <==
################################################################################
################################################################################
################################################################################
################################################################################
############################ BULK GCB PCA & DESEQ2 #############################

## Plan:
## 1. PCA of bulk GCB data 
## 2. Identift TEs upregulated in all portions of the GCB
## 3. Identify TEs specific to LZ, DZ, PB, MB

#################################### SETUP #####################################

library(knitr)
library(tidyverse)
library(matrixStats)
library(data.table)
library(PCAtools)
library(DESeq2)
library(ggplot2)
library(pheatmap)
library(ggsci)
library(ggpubr)
library(edgeR)
library(ashr)
library(cowplot)
library(wesanderson)
library(UpSetR)
library(EnhancedVolcano)

################################### LOAD DATA ##################################

load("r_outputs/02-GCB_Bulk_filt_counts.Rdata")
load("r_outputs/01-metadata.Rdata")
load("r_outputs/01-refs.Rdata")
retro.annot.v2 <- read.csv("/efs/projects/hematological_malignancies_te_analysis/refs/TE_annotation.v2.0.tsv",
                           sep = "\t")
rownames(retro.annot.v2) <- retro.annot.v2$Locus

remove(BL_metadata, DLBCL_metadata, FL_metadata, all_metadata, agirre_metadata)

############################# FUNCTION SCREE PLOT ##############################

pca_standard <- function(tform, metadata, var) {
  
  removeVar <- var
  pca.obj <- PCAtools::pca(assay(tform), 
                           metadata=metadata, 
                           removeVar=removeVar)
  
  cat(sprintf('Removed %d pct low variance variables, %d retained\n', 
              removeVar*100, length(pca.obj$xvars)))
  
  varline <- 50
  varline.x <- min(which(cumsum(pca.obj$variance) >= varline))
  
  horn <- PCAtools::parallelPCA(assay(tform), removeVar = removeVar)
  elbow <- PCAtools::findElbowPoint(pca.obj$variance)
  
  screeplot <-PCAtools::screeplot(pca.obj,
                                  axisLabSize = 6,
                                  components = getComponents(pca.obj, 1:30),
                                  title=paste("Retrotranscriptome SCREE",
                                              metadata$cancer_type[1],
                                              sep=" "),
                                  hline=varline, vline=c(varline.x, horn$n, elbow)
  ) +
    geom_label(aes(x=varline.x+1, y=50, 
                   label = paste0(varline, '% var'), vjust = -1)) +
    geom_label(aes(x = horn$n + 1, y = 50,
                   label = 'Horn\'s', vjust = -1)) +
    geom_label(aes(x = elbow + 1, y = 50,
                   label = 'Elbow method', vjust = -1))
  
  
  cat(sprintf('%d PCs for Elbow method\n', elbow))
  cat(sprintf('%d PCs for Horn method\n', horn$n))
  cat(sprintf('%d PCs needed to explain %d percent of variation\n', 
              varline.x, varline))
  
  return(pca.obj)
}


################################################################################
################################################################################
################################################################################
################################ GCB BULK DESEQ ################################

### DESeq2 (HERVs Only)

bulk.countdat <- GCB_Bulk.filt.herv
cat(sprintf('%d variables\n', nrow(bulk.countdat)))

stopifnot(all(colnames(bulk.countdat) == rownames(bulk_metadata)))

GCB.dds <- DESeq2::DESeqDataSetFromMatrix(countData = bulk.countdat,
                                            colData = bulk_metadata,
                                            design = ~ Cell_type + 0)

GCB.dds <- DESeq2::DESeq(GCB.dds, parallel=T)
GCB.tform <- DESeq2::varianceStabilizingTransformation(GCB.dds, blind=FALSE)

## PCA
GCB.herv.pca.obj <-
  pca_standard(tform = GCB.tform, 
               metadata = bulk_metadata, 
               var = 0.1)

# 3 PCs for Elbow method
# 2 PCs for Horn method
# 1 PCs needed to explain 50 percent of variation

######################## GCB BULK DESEQ (HERVS + GENES) ########################

### DESeq2 (HERVs + Genes)

bulk.g.countdat <- GCB_Bulk.filt.comb
cat(sprintf('%d variables\n', nrow(bulk.g.countdat)))

stopifnot(all(colnames(bulk.g.countdat) == rownames(bulk_metadata)))

GCB.g.dds <- DESeq2::DESeqDataSetFromMatrix(countData = bulk.g.countdat,
                                            colData = bulk_metadata,
                                            design = ~ Cell_type + 0)

GCB.g.dds <- DESeq2::DESeq(GCB.g.dds, parallel=T)
GCB.g.tform <- DESeq2::varianceStabilizingTransformation(GCB.g.dds, blind=FALSE)

## PCA
GCB.g.pca.obj <-
  pca_standard(tform = GCB.g.tform, 
               metadata = bulk_metadata, 
               var = 0.1)

# 3 PCs for Elbow method
# 4 PCs for Horn method
# 1 PCs needed to explain 50 percent of variation

######################### GCB BULK DESEQ (GENES ONLY) ##########################

### DESeq2 (HERVs + Genes)

bulk.gonly.countdat <- GCB_Bulk.filt.tx
cat(sprintf('%d variables\n', nrow(bulk.gonly.countdat)))

stopifnot(all(colnames(bulk.gonly.countdat) == rownames(bulk_metadata)))

GCB.gonly.dds <- DESeq2::DESeqDataSetFromMatrix(countData = bulk.gonly.countdat,
                                            colData = bulk_metadata,
                                            design = ~ Cell_type + 0)

GCB.gonly.dds <- DESeq2::DESeq(GCB.gonly.dds, parallel=T)
GCB.gonly.tform <- DESeq2::varianceStabilizingTransformation(GCB.gonly.dds, blind=FALSE)

## PCA
GCB.gonly.pca.obj <-
  pca_standard(tform = GCB.gonly.tform, 
               metadata = bulk_metadata, 
               var = 0.1)

# 3 PCs for Elbow method
# 3 PCs for Horn method
# 1 PCs needed to explain 50 percent of variation


###################### BULK GCB BIPLOTS HERVs & GENES ##########################

## Biplot with projects (HERVs and genes)

pdf("plots/05e-gcb_pca_genes_hervs.pdf", height=5, width=6)
biplot(GCB.g.pca.obj, 
       lab = NULL,
       showLoadings = FALSE,
       boxedLoadingsNames = TRUE,
       fillBoxedLoadings = alpha("white", 3/4),
       pointSize = 2, 
       encircle = FALSE,
       sizeLoadingsNames = 4,
       lengthLoadingsArrowsFactor = 1.5,
       drawConnectors = TRUE,
       colby = "Cell_type",
       colkey = c("Naive B" = pal_jco("default", alpha = 0.7)(5)[1],
                  "Germinal Center B" = pal_jco("default", alpha = 0.7)(5)[2],
                  "Memory B" = pal_jco("default", alpha = 0.7)(5)[3],
                  "Dark Zone Germinal Center B" = pal_jco("default", alpha = 0.7)(5)[4],
                  "Light Zone Germinal Center B" = pal_jco("default", alpha = 0.7)(5)[5]),
       legendPosition = "right")  +
  theme_cowplot() +
  theme(aspect.ratio = 1)
dev.off()

pdf("plots/05e-gcb_pca_hervs.pdf", height=5, width=6)
## Biplot with projects (HERVs only)
biplot(GCB.herv.pca.obj, 
       lab = NULL,
       showLoadings = FALSE,
       boxedLoadingsNames = TRUE,
       fillBoxedLoadings = alpha("white", 3/4),
       pointSize = 2, 
       encircle = FALSE,
       sizeLoadingsNames = 4,
       lengthLoadingsArrowsFactor = 1.5,
       drawConnectors = TRUE,
       colby = "Cell_type",
       colkey = c("Naive B" = pal_jco("default", alpha = 0.7)(5)[1],
                  "Germinal Center B" = pal_jco("default", alpha = 0.7)(5)[2],
                  "Memory B" = pal_jco("default", alpha = 0.7)(5)[3],
                  "Dark Zone Germinal Center B" = pal_jco("default", alpha = 0.7)(5)[4],
                  "Light Zone Germinal Center B" = pal_jco("default", alpha = 0.7)(5)[5]),
       legendPosition = "right")  +
  theme_cowplot() +
  theme(aspect.ratio = 1)
dev.off()

pdf("plots/05e-gcb_pca_genes.pdf", height=5, width=6)
## Biplot with projects (genes only)
biplot(GCB.gonly.pca.obj, 
       lab = NULL,
       showLoadings = FALSE,
       boxedLoadingsNames = TRUE,
       fillBoxedLoadings = alpha("white", 3/4),
       pointSize = 2, 
       encircle = FALSE,
       sizeLoadingsNames = 4,
       lengthLoadingsArrowsFactor = 1.5,
       drawConnectors = TRUE,
       colby = "Cell_type",
       colkey = c("Naive B" = pal_jco("default", alpha = 0.7)(5)[1],
                  "Germinal Center B" = pal_jco("default", alpha = 0.7)(5)[2],
                  "Memory B" = pal_jco("default", alpha = 0.7)(5)[3],
                  "Dark Zone Germinal Center B" = pal_jco("default", alpha = 0.7)(5)[4],
                  "Light Zone Germinal Center B" = pal_jco("default", alpha = 0.7)(5)[5]),
       legendPosition = "right")  +
  theme_cowplot() +
  theme(aspect.ratio = 1)
dev.off()

################################ SET THRESHOLDS ################################

fc=2 # fold change
l2fc=log2(fc) # log 2 fold change
lfc.cutoff <- 1.5
pval=0.001 # p value threshold

################################################################################
################################################################################
################################################################################
####################### ANALYSIS 1: EACH CELL TYPE VS ALL ######################
################################################################################
################################################################################
################################################################################

############################## TOP GENES & HERVS ###############################

# resultsNames(GCB.g.dds)
# [1] "Cell_typeDark.Zone.Germinal.Center.B"  "Cell_typeGerminal.Center.B"           
# [3] "Cell_typeLight.Zone.Germinal.Center.B" "Cell_typeMemory.B"                    
# [5] "Cell_typeNaive.B"   

gcb_res <- list(
  "DZ" = DESeq2::results(GCB.g.dds, contrast=c(+1, -1/4, -1/4, -1/4, -1/4), alpha=pval),
  "GCB" = DESeq2::results(GCB.g.dds, contrast=c(-1/4, +1, -1/4, -1/4, -1/4), alpha=pval),
  "LZ" = DESeq2::results(GCB.g.dds, contrast=c(-1/4, -1/4, +1, -1/4, -1/4), alpha=pval),
  "MB" = DESeq2::results(GCB.g.dds, contrast=c(-1/4, -1/4, -1/4, +1, -1/4), alpha=pval),
  "NB" = DESeq2::results(GCB.g.dds, contrast=c(-1/4, -1/4, -1/4, -1/4, 1), alpha=pval),
  "DZvLZ" = DESeq2::results(GCB.g.dds, contrast=c("Cell_type", "Dark Zone Germinal Center B", 
                                                  "Light Zone Germinal Center B"), alpha=pval),
  "MBvsNB" = DESeq2::results(GCB.g.dds, contrast=c("Cell_type", "Memory B", "Naive B"), alpha=pval)
)

gcb_res <- lapply(gcb_res, function(r) {
  r$display <- gene_table[rownames(r),]$display
  r$class <- gene_table[rownames(r),]$gene_type
  r
})

sig <- lapply(gcb_res, function(r) {
  s <- subset(r, padj < pval & abs(log2FoldChange) > lfc.cutoff)
  s[order(s$padj),]
})

for (n in names(sig)) {
  cat("\n#--- Contrast", n, "---#\n")
  summary(sig[[n]])
}

########################## TOP GENES & HERVS NO GCB ############################

gcb_res_no_gcb <- list(
  "DZ" = DESeq2::results(GCB.g.dds, contrast=c(+1, 0, -1/3, -1/3, -1/3), alpha=pval),
  "LZ" = DESeq2::results(GCB.g.dds, contrast=c(-1/3, 0, +1, -1/3, -1/3), alpha=pval),
  "MB" = DESeq2::results(GCB.g.dds, contrast=c(-1/3, 0, -1/3, +1, -1/3), alpha=pval),
  "NB" = DESeq2::results(GCB.g.dds, contrast=c(-1/3, 0, -1/3, -1/3, 1), alpha=pval),
  "DZvLZ" = DESeq2::results(GCB.g.dds, contrast=c("Cell_type", "Dark Zone Germinal Center B", 
                                                  "Light Zone Germinal Center B"), alpha=pval),
  "MBvsNB" = DESeq2::results(GCB.g.dds, contrast=c("Cell_type", "Memory B", "Naive B"), alpha=pval)
)

gcb_res_no_gcb <- lapply(gcb_res_no_gcb, function(r) {
  r$display <- gene_table[rownames(r),]$display
  r$class <- gene_table[rownames(r),]$gene_type
  r
})

sig_no_gcb <- lapply(gcb_res_no_gcb, function(r) {
  s <- subset(r, padj < pval & abs(log2FoldChange) > lfc.cutoff)
  s[order(s$padj),]
})

for (n in names(sig_no_gcb)) {
  cat("\n#--- Contrast", n, "---#\n")
  summary(sig_no_gcb[[n]])
}


############################### TOP HERVs ONLY #################################

gcb_res_herv <- list(
  "DZ" = DESeq2::results(GCB.dds, contrast=c(+1, -1/4, -1/4, -1/4, -1/4), alpha=pval),
  "GCB" = DESeq2::results(GCB.dds, contrast=c(-1/4, +1, -1/4, -1/4, -1/4), alpha=pval),
  "LZ" = DESeq2::results(GCB.dds, contrast=c(-1/4, -1/4, +1, -1/4, -1/4), alpha=pval),
  "MB" = DESeq2::results(GCB.dds, contrast=c(-1/4, -1/4, -1/4, +1, -1/4), alpha=pval),
  "NB" = DESeq2::results(GCB.dds, contrast=c(-1/4, -1/4, -1/4, -1/4, 1), alpha=pval),
  "DZvLZ" = DESeq2::results(GCB.dds, contrast=c("Cell_type", "Dark Zone Germinal Center B", 
                                                "Light Zone Germinal Center B"), alpha=pval),
  "MBvsNB" = DESeq2::results(GCB.dds, contrast=c("Cell_type", "Memory B", "Naive B"), alpha=pval)
)

gcb_res_herv <- lapply(gcb_res_herv, function(r) {
  r$display <- gene_table[rownames(r),]$display
  r$class <- gene_table[rownames(r),]$gene_type
  r$TE_type <- retro.annot.v2[rownames(r),]$TE_type
  r
})


sig_herv <- lapply(gcb_res_herv, function(r) {
  s <- subset(r, padj < pval & abs(log2FoldChange) > lfc.cutoff)
  s[order(s$padj),]
})

sink(file = "r_outputs/05e-gcb_herv_deseq.txt")
for (n in names(sig_herv)) {
  cat("\n#--- Contrast", n, "---#\n")
  summary(sig_herv[[n]])
}
sink(file = NULL)

sink(file = "r_outputs/05e-gcb_herv_deseq_top5.txt")
for (n in names(sig_herv)) {
  cat("\n#--- Contrast", n, "---#\n")
  print(head(sig_herv[[n]]))
}
sink(file = NULL)

############################ TOP HERVs ONLY NO GCB #############################

gcb_res_herv_no_gcb <- list(
  "DZ" = DESeq2::results(GCB.dds, contrast=c(+1, 0, -1/3, -1/3, -1/3), alpha=pval),
  "LZ" = DESeq2::results(GCB.dds, contrast=c(-1/3, 0, +1, -1/3, -1/3), alpha=pval),
  "MB" = DESeq2::results(GCB.dds, contrast=c(-1/3, 0, -1/3, +1, -1/3), alpha=pval),
  "NB" = DESeq2::results(GCB.dds, contrast=c(-1/3, 0, -1/3, -1/3, 1), alpha=pval),
  "DZvLZ" = DESeq2::results(GCB.dds, contrast=c("Cell_type", "Dark Zone Germinal Center B", 
                                                  "Light Zone Germinal Center B"), alpha=pval),
  "MBvsNB" = DESeq2::results(GCB.dds, contrast=c("Cell_type", "Memory B", "Naive B"), alpha=pval)
)

gcb_res_herv_no_gcb <- lapply(gcb_res_herv_no_gcb, function(r) {
  r$display <- gene_table[rownames(r),]$display
  r$class <- gene_table[rownames(r),]$gene_type
  r$TE_type <- retro.annot.v2[rownames(r),]$TE_type
  r
})


sig_herv_no_gcb <- lapply(gcb_res_herv_no_gcb, function(r) {
  s <- subset(r, padj < pval & abs(log2FoldChange) > lfc.cutoff)
  s[order(s$padj),]
})

sink(file = "r_outputs/05e-gcb_herv_no_gcb_deseq.txt")
for (n in names(sig_herv_no_gcb)) {
  cat("\n#--- Contrast", n, "---#\n")
  summary(sig_herv_no_gcb[[n]])
}
sink(file = NULL)

sink(file = "r_outputs/05e-gcb_herv_no_gcb_deseq_top5.txt")
for (n in names(sig_herv_no_gcb)) {
  cat("\n#--- Contrast", n, "---#\n")
  print(head(sig_herv_no_gcb[[n]]))
}
sink(file = NULL)


############################ UPSET GENES & HERVS ###############################

upvars <- lapply(sig[1:5], function(r) rownames(subset(r, log2FoldChange>0)))
downvars <- lapply(sig[1:5], function(r) rownames(subset(r, log2FoldChange<0)))

pdf("plots/05e-gcb_upset_upvars.pdf", height=5, width=7)
upset(fromList(upvars), sets=c("DZ", "GCB", "LZ", "MB", "NB"),  
      keep.order = T, order.by='degree', decreasing=F,
      text.scale = c(2, 2, 1.5, 1.5, 1.5, 1.5))
dev.off()

pdf("plots/05e-gcb_upset_dnwars.pdf", height=5, width=7)
upset(fromList(downvars), sets=c("DZ", "GCB", "LZ", "MB", "NB"),  
      keep.order = T, order.by='degree', decreasing=F,
      text.scale = c(2, 2, 1.5, 1.5, 1.5, 1.5))
dev.off()

up.binmat <- fromList(upvars)
rn <- do.call(c, upvars)
rn <- rn[!duplicated(rn)]
rownames(up.binmat) <- rn
rm(rn)

dn.binmat <- fromList(downvars)
rn <- do.call(c, downvars)
rn <- rn[!duplicated(rn)]
rownames(dn.binmat) <- rn
rm(rn)

######################## UPSET GENES & HERVS NO GCB ############################

upvars_no_gcb <- lapply(sig_no_gcb[1:4], function(r) 
  rownames(subset(r, log2FoldChange>0)))
downvars_no_gcb <- lapply(sig_no_gcb[1:5], function(r) 
  rownames(subset(r, log2FoldChange<0)))

pdf("plots/05e-gcb_no_gcb_upset_upvars.pdf", height=5, width=7)
upset(fromList(upvars_no_gcb), sets=c("DZ", "LZ", "MB", "NB"),  
      keep.order = T, order.by='degree', decreasing=F,
      text.scale = c(2, 2, 1.5, 1.5, 1.5, 1.5))
dev.off()

pdf("plots/05e-gcb_no_gcb_upset_dnwars.pdf", height=5, width=7)
upset(fromList(downvars_no_gcb), sets=c("DZ","LZ", "MB", "NB"),  
      keep.order = T, order.by='degree', decreasing=F,
      text.scale = c(2, 2, 1.5, 1.5, 1.5, 1.5))
dev.off()

up.no.gcb.binmat <- fromList(upvars_no_gcb)
rn <- do.call(c, upvars_no_gcb)
rn <- rn[!duplicated(rn)]
rownames(up.no.gcb.binmat) <- rn
rm(rn)

dn.no.gcb.binmat <- fromList(downvars_no_gcb)
rn <- do.call(c, downvars_no_gcb)
rn <- rn[!duplicated(rn)]
rownames(dn.no.gcb.binmat) <- rn
rm(rn)

############################# UPSET HERVs ONLY #################################

upvars_hervs <- lapply(sig_herv[1:5], function(r) rownames(subset(r, log2FoldChange>0)))
downvars_hervs <- lapply(sig_herv[1:5], function(r) rownames(subset(r, log2FoldChange<0)))

pdf("plots/05e-gcb_upset_upvars_hervs.pdf", height=5, width=7)
upset(fromList(upvars_hervs), sets=c("DZ", "GCB", "LZ", "MB", "NB"),  
      keep.order = T, order.by='degree', decreasing=F,
      text.scale = c(1.5, 2, 1.5, 1.5, 1.5, 1.5))
dev.off()

pdf("plots/05e-gcb_upset_dnvars_hervs.pdf", height=5, width=7)
upset(fromList(downvars_hervs), sets=c("DZ", "GCB", "LZ", "MB", "NB"),  
      keep.order = T, order.by='degree', decreasing=F,
      text.scale = c(1.5, 2, 1.5, 1.5, 1.5, 1.5))
dev.off()

up.binmat.hervs <- fromList(upvars_hervs)
rn <- do.call(c, upvars_hervs)
rn <- rn[!duplicated(rn)]
rownames(up.binmat.hervs) <- rn
rm(rn)

dn.binmat.hervs <- fromList(downvars_hervs)
rn <- do.call(c, downvars_hervs)
rn <- rn[!duplicated(rn)]
rownames(dn.binmat.hervs) <- rn
rm(rn)

######################### UPSET HERVs ONLY NO GCB ##############################

upvars_hervs_no_gcb <- lapply(sig_herv_no_gcb[1:5], function(r) 
  rownames(subset(r, log2FoldChange>0)))
downvars_hervs_no_gcb <- lapply(sig_herv_no_gcb[1:5],
                                function(r) rownames(subset(r, log2FoldChange<0)))

pdf("plots/05e-gcb_no_gcb_upset_upvars_hervs.pdf", height=5, width=7)
upset(fromList(upvars_hervs_no_gcb), sets=c("DZ", "LZ", "MB", "NB"),  
      keep.order = T, order.by='degree', decreasing=F,
      text.scale = c(1.5, 2, 1.5, 1.5, 1.5, 1.5))
dev.off()

pdf("plots/05e-gcb_no_gcb_upset_dnvars_hervs.pdf", height=5, width=7)
upset(fromList(downvars_hervs_no_gcb), sets=c("DZ", "LZ", "MB", "NB"),  
      keep.order = T, order.by='degree', decreasing=F,
      text.scale = c(1.5, 2, 1.5, 1.5, 1.5, 1.5))
dev.off()

up.no.gcb.binmat.hervs <- fromList(upvars_hervs_no_gcb)
rn <- do.call(c, upvars_hervs_no_gcb)
rn <- rn[!duplicated(rn)]
rownames(up.no.gcb.binmat.hervs) <- rn
rm(rn)

dn.no.gcb.binmat.hervs <- fromList(downvars_hervs_no_gcb)
rn <- do.call(c, downvars_hervs_no_gcb)
rn <- rn[!duplicated(rn)]
rownames(dn.no.gcb.binmat.hervs) <- rn
rm(rn)

################################# HEATMAPS #####################################

################################### SETUP ######################################

# Colors for cells
cols <- rgb_gsea(palette = c("default"), n = 14, alpha = 0.7, reverse = FALSE)

# Annotation column
df <- as.data.frame(colData(GCB.dds)[,c("BioSample","Cell_type")])
df <- subset(df, select = -c(1))

# Create colors for each group
annoCol <-  pal_jco("default", alpha = 0.7)(5)
annoCol2 <- pal_aaas("default", alpha=0.7)(3)
names(annoCol2) <- unique(retro.annot.v2$TE_type)
names(annoCol) <- unique(df$Cell_type)
annoCol <- list(Cell_type = annoCol, TE_type = annoCol2)

######################### UPREGULATED IN ALL GROUPS ############################

top.genes.hervs <- rownames(up.binmat)
top.hervs <- rownames(up.binmat.hervs)


annoRow <- as.data.frame(retro.annot.v2[,c("TE_type", "Locus")])
annoRow <- annoRow[top.hervs,]
annoRow <- subset(annoRow, select = -c(2))

pdf("plots/05e-gcb_top_hervs_upregulated_all.pdf", height=10, width=10)
pheatmap(assay(GCB.tform)[top.hervs,], 
         main="Upregulated HERVs, all clusters",
         cluster_rows=TRUE,
         show_rownames=FALSE,
         show_colnames = FALSE,
         color = cols,
         scale="row",
         breaks=seq(-3,3,length.out=14),
         labels_row = gene_table[top.hervs,]$display,
         cluster_cols=TRUE, 
         treeheight_row=0,
         annotation_col=df,
         annotation_row=annoRow,
         annotation_colors = annoCol)
dev.off()

pdf("plots/05e-gcb_top_hervs_genes_upregulated_all.pdf", height=10, width=10)
pheatmap(assay(GCB.g.tform)[top.genes.hervs,],
         main="Upregulated Genes and HERVs, all clusters",
         cluster_rows=TRUE,
         show_rownames=FALSE,
         show_colnames = FALSE,
         color = cols,
         scale="row",
         breaks=seq(-3,3,length.out=14),
         labels_row = gene_table[top.genes.hervs,]$display,
         cluster_cols=TRUE, 
         treeheight_row=0,
         annotation_col=df,
         annotation_colors = annoCol)
dev.off()


########################### DE HERVs PER GC SITE ###############################


makeheatmap <- function(topgenes, ...) {
  args <- list(...)
  mat <- assay(GCB.tform)[unique(topgenes), ]
  rowdist <- as.dist((1 - cor(t(mat), method='spearman'))/2)
  
  annotation_col <- df
  cols <- rgb_gsea(palette = c("default"), n = 14, alpha = 0.7, reverse = FALSE)
  
  annoRow <- as.data.frame(retro.annot.v2[,c("TE_type", "Locus")])
  annoRow <- annoRow[topgenes,]
  annoRow <- subset(annoRow, select = -c(2))
  
  pheatmap(mat,
           color=cols,
           scale="row", breaks=seq(-3,3,length.out=14),
           clustering_distance_rows = rowdist,
           clustering_method="average",
           annotation_col = df,
           annotation_colors = annoCol,
           annotation_row = annoRow,
           show_colnames=F, 
           show_rownames = T,
           fontsize = 8, fontsize_row = 6, border_color = NA,
           legend=T,
           treeheight_col=10,
           treeheight_row=10, 
           cutree_cols=4,
           ...
  )
}


for(clust in c("DZ", "GCB", "LZ", "MB", "NB")) {
  tg <- rownames(sig_herv[[clust]][1:75,])
  p <- makeheatmap(tg, main=paste0('DE in cluster ', clust))
  pdf(paste0("plots/05e-gcb_top_de_hervs_", clust, ".pdf"), height=7, width=7)
  print(p)
  dev.off()
}

############################# VOLCANO DZ VS LZ #################################

pdf("plots/05e-gcb_volcano_DZ_v_all.pdf", height=8, width=8)
EnhancedVolcano(sig_herv$DZ,
                lab = rownames(sig_herv$DZ),
                x = 'log2FoldChange',
                y = 'pvalue',
                title = 'All vs DZ')
dev.off()

pdf("plots/05e-gcb_volcano_LZ_v_all.pdf", height=8, width=8)
EnhancedVolcano(sig_herv$LZ,
                lab = rownames(sig_herv$LZ),
                x = 'log2FoldChange',
                y = 'pvalue',
                title = 'All vs LZ')
dev.off()

pdf("plots/05e-gcb_volcano_NB_v_all.pdf", height=8, width=8)
EnhancedVolcano(sig_herv$NB,
                lab = rownames(sig_herv$NB),
                x = 'log2FoldChange',
                y = 'pvalue',
                title = 'All vs NB')
dev.off()

pdf("plots/05e-gcb_volcano_MB_v_all.pdf", height=8, width=8)
EnhancedVolcano(sig_herv$MB,
                lab = rownames(sig_herv$MB),
                x = 'log2FoldChange',
                y = 'pvalue',
                title = 'All vs MB')
dev.off()

pdf("plots/05e-gcb_volcano_DZ_v_LZ.pdf", height=8, width=8)
EnhancedVolcano(sig_herv$DZvLZ,
                lab = rownames(sig_herv$DZvLZ),
                x = 'log2FoldChange',
                y = 'pvalue',
                title = 'LZ vs DZ')
dev.off()


pdf("plots/05e-gcb_volcano_NB_v_MB.pdf", height=8, width=8)
EnhancedVolcano(sig_herv$MBvsNB,
                lab = rownames(sig_herv$MBvsNB),
                x = 'log2FoldChange',
                y = 'pvalue',
                title = 'NB vs MB')
dev.off()
########################## PLOT INDIVIDUAL HERVs ###############################

plot.counts <- function(df, gene) {
  
  as.data.frame(plotCounts(df, 
                           gene=gene, 
                           intgroup="Cell_type", 
                           returnData = TRUE)) %>%
    ggplot(aes(x=Cell_type, y=count, fill=Cell_type))  +
    geom_boxplot() +
    theme_pubr() +
    theme(legend.position="none",
          axis.text.x = element_text(angle=45, hjust=1)) +
    xlab("Cell Type") +
    ylab("Counts") +
    scale_x_discrete(labels=c("Dark Zone Germinal Center B" = "DZ",
                              "Germinal Center B" = "GCB",
                              "Light Zone Germinal Center B" = "LZ",
                              "Memory B" = "MB",
                              "Naive B" = "NB")) +
    scale_fill_manual(values = c("Naive B" = pal_jco("default", alpha = 0.7)(5)[1],
                                 "Germinal Center B" = pal_jco("default", alpha = 0.7)(5)[2],
                                 "Memory B" = pal_jco("default", alpha = 0.7)(5)[3],
                                 "Dark Zone Germinal Center B" = pal_jco("default", alpha = 0.7)(5)[4],
                                 "Light Zone Germinal Center B" = pal_jco("default", alpha = 0.7)(5)[5])) + 
    scale_y_continuous(expand = c(0, 0), limits = c(0, NA)) +
    ggtitle(gene) + 
    theme(plot.title = element_text(hjust = 0.5),
          aspect.ratio = 1)
}


p1 <- plot.counts(GCB.g.dds, "HML2_3q12.3")
p2 <- plot.counts(GCB.g.dds, "ERV316A3_6p21.33c")
p3 <- plot.counts(GCB.g.dds, "MER4_17q21.2d")
p4 <- plot.counts(GCB.g.dds, "HML3_5p15.33d")
p5 <- plot.counts(GCB.g.dds, "HERVEA_5q22.2")
p6 <- plot.counts(GCB.g.dds, "HERVH_7q11.21")
p7 <- plot.counts(GCB.g.dds, "HERVH_1q32.1b")
p8 <- plot.counts(GCB.g.dds, "HERVW_1q32.1")
p9 <- plot.counts(GCB.g.dds, "HML5_1q22")

pdf("plots/05e-gcb_individual_hervs.pdf", height=12, width=12)
plot_grid(p1, p2, p3, p4, p5, p6, p7, p8, p9,
          nrow = 3, 
          ncol = 3,
          labels = "AUTO")
dev.off()

pdf("plots/05e-gcb_healthy_HARLEQUIN_17q21.31.pdf", height=3, width=3)
plot.counts(GCB.g.dds, "HARLEQUIN_17q21.31")
dev.off()

pdf("plots/05e-gcb_healthy_HARLEQUIN_1q32.1.pdf", height=3, width=3)
plot.counts(GCB.g.dds, "HARLEQUIN_1q32.1")
dev.off()

pdf("plots/05e-gcb_healthy_HML5_1q22.pdf", height=3, width=3)
plot.counts(GCB.g.dds, "HML5_1q22")
dev.off()

pdf("plots/05e-gcb_healthy_LILRB1.pdf", height=3, width=3)
plot.counts(GCB.g.dds, "ENSG00000104972.16")
dev.off()

############################### FAMILY LEVEL UP ################################

upreg_hervs_no_gcb_df <- do.call(rbind, lapply(upvars_hervs_no_gcb, data.frame))
colnames(upreg_hervs_no_gcb_df) <- c("herv")
upreg_hervs_no_gcb_df$cell_type <- rownames(upreg_hervs_no_gcb_df)
upreg_hervs_no_gcb_df$cell_type <- gsub("\\..*","",upreg_hervs_no_gcb_df$cell_type)
upreg_hervs_no_gcb_df$family <- retro.annot$family[match(upreg_hervs_no_gcb_df$herv, 
                                                         retro.annot$locus)]

upreg_families_no_gcb <-
  upreg_hervs_no_gcb_df %>% dplyr::count(family, cell_type, sort = TRUE) 

upreg_families_no_gcb<-upreg_families_no_gcb[!(upreg_families_no_gcb$cell_type=="DZvLZ"),]

pdf("plots/05e-gcb_upreg_families.pdf", height=8, width=6)
ggplot(upreg_families_no_gcb, aes(fill=reorder(family, -n), y=cell_type, x=n)) + 
  geom_bar(position="fill", stat="identity", colour="black", size=0.3) + 
  scale_fill_manual(values = c(pal_futurama("planetexpress")(12), 
                               pal_npg("nrc", alpha = 0.7)(10),
                               pal_jco("default", alpha=0.7)(10),
                               pal_nejm("default", alpha=0.7)(8),
                               pal_tron("legacy", alpha=0.7)(7),
                               pal_lancet("lanonc", alpha=0.7)(9),
                               pal_startrek("uniform", alpha=0.7)(7)),
                    breaks = unique(retro.annot$family),
                    labels = unique(retro.annot$family)) + 
  
  coord_flip() +
  theme_pubclean() +  
  guides(fill=guide_legend(title="TE Family")) +
  ylab("Upregulated HERVs by Family") +
  xlab("Proportion of HERVs") + 
  theme(legend.position = c("right")) + 
  guides(fill = guide_legend(title = "HERV family", ncol = 2))
dev.off()

############################### FAMILY LEVEL DOWN ##############################

down_hervs_no_gcb_df <- do.call(rbind, lapply(downvars_hervs_no_gcb, data.frame))
colnames(down_hervs_no_gcb_df) <- c("herv")
down_hervs_no_gcb_df$cell_type <- rownames(down_hervs_no_gcb_df)
down_hervs_no_gcb_df$cell_type <- gsub("\\..*","",down_hervs_no_gcb_df$cell_type)
down_hervs_no_gcb_df$family <- retro.annot$family[match(down_hervs_no_gcb_df$herv, 
                                                         retro.annot$locus)]

downreg_families_no_gcb <-
  down_hervs_no_gcb_df %>% dplyr::count(family, cell_type, sort = TRUE) 

downreg_families_no_gcb<-downreg_families_no_gcb[!(downreg_families_no_gcb$cell_type=="DZvLZ"),]

pdf("plots/05e-gcb_downreg_families.pdf", height=8, width=6)
ggplot(downreg_families_no_gcb, aes(fill=reorder(family, -n), y=cell_type, x=n)) + 
  geom_bar(position="fill", stat="identity", colour="black", size=0.3) + 
  scale_fill_manual(values = c(pal_futurama("planetexpress")(12), 
                               pal_npg("nrc", alpha = 0.7)(10),
                               pal_jco("default", alpha=0.7)(10),
                               pal_nejm("default", alpha=0.7)(8),
                               pal_tron("legacy", alpha=0.7)(7),
                               pal_lancet("lanonc", alpha=0.7)(9),
                               pal_startrek("uniform", alpha=0.7)(7)),
                    breaks = unique(retro.annot$family),
                    labels = unique(retro.annot$family)) + 
  
  coord_flip() +
  theme_pubclean() +  
  guides(fill=guide_legend(title="TE Family")) +
  ylab("Downregulated HERVs by Family") +
  xlab("Proportion of HERVs") + 
  theme(legend.position = c("right")) + 
  guides(fill = guide_legend(title = "HERV family", ncol = 2))
dev.off()


############################# MERGED FAMILY UP/DOWN ############################


updown_family <- do.call(rbind, (list(Upregulated = upreg_families_no_gcb, 
                                      Dowregulated = downreg_families_no_gcb)))
updown_family$expression <- rownames(updown_family)
updown_family$expression <- gsub("\\..*","",updown_family$expression)
rownames(updown_family)<-NULL

pdf("plots/05e-gcb_updown_families.pdf", height=8, width=10)
ggplot(updown_family, aes(fill=reorder(family, -n), y=cell_type, x=n)) + 
  geom_bar(position="fill", stat="identity", colour="black", size=0.3) + 
  scale_fill_manual(values = c(pal_futurama("planetexpress")(12), 
                               pal_npg("nrc", alpha = 0.7)(10),
                               pal_jco("default", alpha=0.7)(10),
                               pal_nejm("default", alpha=0.7)(8),
                               pal_tron("legacy", alpha=0.7)(7),
                               pal_lancet("lanonc", alpha=0.7)(9),
                               pal_startrek("uniform", alpha=0.7)(7)),
                    breaks = unique(retro.annot$family),
                    labels = unique(retro.annot$family)) + 
  
  coord_flip() +
  theme_pubclean() +  
  guides(fill=guide_legend(title="TE Family")) +
  ylab("Downregulated HERVs by Family") +
  xlab("Proportion of HERVs") + 
  theme(legend.position = c("right")) + 
  guides(fill = guide_legend(title = "HERV family", ncol = 2)) +
  facet_wrap(~ expression, ncol = 2)
dev.off()

#################################### TE TYPE ###################################

rownames(upreg_hervs_no_gcb_df) <- NULL
upreg_hervs_no_gcb_df$type <- retro.annot.v2$TE_type[
  match(upreg_hervs_no_gcb_df$herv, retro.annot.v2$Locus)]

upreg_hervs_no_gcb_df<-upreg_hervs_no_gcb_df[!(upreg_hervs_no_gcb_df$cell_type=="DZvLZ"),]

upreg_te_type <-
  upreg_hervs_no_gcb_df %>% dplyr::count(type, cell_type, sort = TRUE) 

pdf("plots/05e-gcb_upreg_hervs_type.pdf", height=4, width=4)
ggplot(upreg_te_type, aes(fill=reorder(type, -n), y=cell_type, x=n)) + 
  geom_bar(position="stack", stat="identity", colour="black", size=0.3) + 
  scale_fill_manual(values = c(pal_aaas("default", alpha=0.7)(3))) + 
  coord_flip() +
  theme_pubclean() +  
  guides(fill=guide_legend(title="HERV Type")) +
  ylab("Upregulated HERVs by Type") +
  xlab("Proportion of HERVs") + 
  theme(legend.position = c("right")) +
  theme(aspect.ratio = 1)
dev.off()

rownames(down_hervs_no_gcb_df) <- NULL
down_hervs_no_gcb_df$type <- retro.annot.v2$TE_type[
  match(down_hervs_no_gcb_df$herv, retro.annot.v2$Locus)]

down_hervs_no_gcb_df<-down_hervs_no_gcb_df[!(down_hervs_no_gcb_df$cell_type=="DZvLZ"),]

downreg_te_type <-
  down_hervs_no_gcb_df %>% dplyr::count(type, cell_type, sort = TRUE) 

pdf("plots/05e-gcb_downreg_hervs_type.pdf", height=4, width=4)
ggplot(downreg_te_type, aes(fill=reorder(type, -n), y=cell_type, x=n)) + 
  geom_bar(position="stack", stat="identity", colour="black", size=0.3) + 
  scale_fill_manual(values = c(pal_aaas("default", alpha=0.7)(3))) + 
  coord_flip() +
  theme_pubclean() +  
  guides(fill=guide_legend(title="HERV Type")) +
  ylab("Downregulated HERVs by Type") +
  xlab("Proportion of HERVs") + 
  theme(legend.position = c("right")) +
  theme(aspect.ratio = 1)
dev.off()

################################################################################
################################################################################
################################################################################
######################## ANALYSIS 2: NAIVE AS BASELINE #########################
################################################################################
################################################################################
################################################################################

gcb_res_naive <- list(
  "DZ" = DESeq2::results(GCB.g.dds,contrast=c("Cell_type", "Dark Zone Germinal Center B", 
                                              "Naive B"), alpha=pval),
  "GCB" = DESeq2::results(GCB.g.dds,contrast=c("Cell_type", "Germinal Center B", 
                                               "Naive B"), alpha=pval),
  "LZ" = DESeq2::results(GCB.g.dds, contrast=c("Cell_type", "Light Zone Germinal Center B", 
                                               "Naive B"), alpha=pval),
  "MB" = DESeq2::results(GCB.g.dds, contrast=c("Cell_type", "Memory B", 
                                               "Naive B"), alpha=pval),
  "NB" = DESeq2::results(GCB.g.dds, contrast=c(-1/4, -1/4, -1/4, -1/4, 1), alpha=pval))

gcb_res_naive <- lapply(gcb_res_naive, function(r) {
  r$display <- gene_table[rownames(r),]$display
  r$class <- gene_table[rownames(r),]$gene_type
  r
})

sig_naive <- lapply(gcb_res_naive, function(r) {
  s <- subset(r, padj < pval & abs(log2FoldChange) > lfc.cutoff)
  s[order(s$padj),]
})

for (n in names(sig_naive)) {
  cat("\n#--- Contrast", n, "---#\n")
  summary(sig_naive[[n]])
}

############################### TOP HERVs ONLY #################################

gcb_res_naive_herv <- list(
  "DZ" = DESeq2::results(GCB.dds,contrast=c("Cell_type", "Dark Zone Germinal Center B", 
                                              "Naive B"), alpha=pval),
  "GCB" = DESeq2::results(GCB.dds,contrast=c("Cell_type", "Germinal Center B", 
                                               "Naive B"), alpha=pval),
  "LZ" = DESeq2::results(GCB.dds, contrast=c("Cell_type", "Light Zone Germinal Center B", 
                                               "Naive B"), alpha=pval),
  "MB" = DESeq2::results(GCB.dds, contrast=c("Cell_type", "Memory B", 
                                               "Naive B"), alpha=pval),
  "NB" = DESeq2::results(GCB.dds, contrast=c(-1/4, -1/4, -1/4, -1/4, 1), alpha=pval))

gcb_res_naive_herv <- lapply(gcb_res_naive_herv, function(r) {
  r$display <- gene_table[rownames(r),]$display
  r$class <- gene_table[rownames(r),]$gene_type
  r$TE_type <- retro.annot.v2[rownames(r),]$TE_type
  r
})


sig_naive_herv <- lapply(gcb_res_naive_herv, function(r) {
  s <- subset(r, padj < pval & abs(log2FoldChange) > lfc.cutoff)
  s[order(s$padj),]
})

sink(file = "r_outputs/05e-gcb_naive_herv_deseq.txt")
for (n in names(sig_naive_herv)) {
  cat("\n#--- Contrast", n, "---#\n")
  summary(sig_naive_herv[[n]])
}
sink(file = NULL)

sink(file = "r_outputs/05e-gcb_naive_herv_deseq_top5.txt")
for (n in names(sig_naive_herv)) {
  cat("\n#--- Contrast", n, "---#\n")
  print(head(sig_naive_herv[[n]]))
}
sink(file = NULL)

############################ UPSET GENES & HERVS ###############################

upvars_naive <- lapply(sig_naive[1:5], function(r) rownames(subset(r, log2FoldChange>0)))
downvars_naive <- lapply(sig_naive[1:5], function(r) rownames(subset(r, log2FoldChange<0)))

pdf("plots/05e-gcb_naive_upset_upvars.pdf", height=5, width=7)
upset(fromList(upvars_naive), sets=c("DZ", "GCB", "LZ", "MB", "NB"),  
      keep.order = T, order.by='degree', decreasing=F,
      text.scale = c(2, 2, 1.5, 1.5, 1.5, 1.5))
dev.off()

pdf("plots/05e-gcb_naive_upset_dnwars.pdf", height=5, width=7)
upset(fromList(downvars_naive), sets=c("DZ", "GCB", "LZ", "MB", "NB"),  
      keep.order = T, order.by='degree', decreasing=F,
      text.scale = c(2, 2, 1.5, 1.5, 1.5, 1.5))
dev.off()

up.naive.binmat <- fromList(upvars_naive)
rn <- do.call(c, upvars_naive)
rn <- rn[!duplicated(rn)]
rownames(up.naive.binmat) <- rn
rm(rn)

dn.naive.binmat <- fromList(downvars_naive)
rn <- do.call(c, downvars_naive)
rn <- rn[!duplicated(rn)]
rownames(dn.naive.binmat) <- rn
rm(rn)

############################# UPSET HERVs ONLY #################################

upvars_naive_hervs <- lapply(sig_naive_herv[1:5], function(r) rownames(subset(r, log2FoldChange>0)))
downvars_naive_hervs <- lapply(sig_naive_herv[1:5], function(r) rownames(subset(r, log2FoldChange<0)))

pdf("plots/05e-gcb_naive_upset_upvars_hervs.pdf", height=5, width=7)
upset(fromList(upvars_naive_hervs), sets=c("DZ", "GCB", "LZ", "MB", "NB"),  
      keep.order = T, order.by='degree', decreasing=F,
      text.scale = c(1.5, 2, 1.5, 1.5, 1.5, 1.5))
dev.off()

pdf("plots/05e-gcb_naive_upset_dnvars_hervs.pdf", height=5, width=7)
upset(fromList(downvars_naive_hervs), sets=c("DZ", "GCB", "LZ", "MB", "NB"),  
      keep.order = T, order.by='degree', decreasing=F,
      text.scale = c(1.5, 2, 1.5, 1.5, 1.5, 1.5))
dev.off()

up.naive.binmat.hervs <- fromList(upvars_naive_hervs)
rn <- do.call(c, upvars_naive_hervs)
rn <- rn[!duplicated(rn)]
rownames(up.naive.binmat.hervs) <- rn
rm(rn)

dn.naive.binmat.hervs <- fromList(downvars_naive_hervs)
rn <- do.call(c, downvars_naive_hervs)
rn <- rn[!duplicated(rn)]
rownames(dn.naive.binmat.hervs) <- rn
rm(rn)

##################### UPREGULATED IN ALL GROUPS VS NAIVE #######################

top.genes.hervs <- rownames(up.naive.binmat)
top.hervs <- rownames(up.naive.binmat.hervs)


annoRow <- as.data.frame(retro.annot.v2[,c("TE_type", "Locus")])
annoRow <- annoRow[top.hervs,]
annoRow <- subset(annoRow, select = -c(2))

pdf("plots/05e-gcb_naive_top_hervs_upregulated_all.pdf", height=10, width=10)
pheatmap(assay(GCB.tform)[top.hervs,], 
         main="Upregulated HERVs, all clusters",
         cluster_rows=TRUE,
         show_rownames=FALSE,
         show_colnames = FALSE,
         color = cols,
         scale="row",
         breaks=seq(-3,3,length.out=14),
         labels_row = gene_table[top.hervs,]$display,
         cluster_cols=TRUE, 
         treeheight_row=0,
         annotation_col=df,
         annotation_row=annoRow,
         annotation_colors = annoCol)
dev.off()

pdf("plots/05e-gcb_naive_top_hervs_genes_upregulated_all.pdf", height=10, width=10)
pheatmap(assay(GCB.g.tform)[top.genes.hervs,],
         main="Upregulated Genes and HERVs, all clusters",
         cluster_rows=TRUE,
         show_rownames=FALSE,
         show_colnames = FALSE,
         color = cols,
         scale="row",
         breaks=seq(-3,3,length.out=14),
         labels_row = gene_table[top.genes.hervs,]$display,
         cluster_cols=TRUE, 
         treeheight_row=0,
         annotation_col=df,
         annotation_colors = annoCol)
dev.off()

for(clust in c("DZ", "GCB", "LZ", "NB")) {
  tg <- rownames(sig_naive_herv[[clust]][1:75,])
  p <- makeheatmap(tg, main=paste0('DE in cluster ', clust))
  pdf(paste0("plots/05e-gcb_naive_top_de_hervs_", clust, ".pdf"), height=7, width=7)
  print(p)
  dev.off()
}

for(clust in c("MB")) {
  tg <- rownames(sig_naive_herv[[clust]][1:25,])
  p <- makeheatmap(tg, main=paste0('DE in cluster ', clust))
  pdf(paste0("plots/05e-gcb_naive_top_de_hervs_", clust, ".pdf"), height=7, width=7)
  print(p)
  dev.off()
}

############################# VOLCANO DZ VS LZ #################################

pdf("plots/05e-gcb_naive_volcano_DZ_v_NB.pdf", height=8, width=8)
EnhancedVolcano(sig_naive_herv$DZ,
                lab = rownames(sig_naive_herv$DZ),
                x = 'log2FoldChange',
                y = 'pvalue',
                title = 'NB vs DZ')
dev.off()

pdf("plots/05e-gcb_naive_volcano_LZ_v_NB.pdf", height=8, width=8)
EnhancedVolcano(sig_naive_herv$LZ,
                lab = rownames(sig_naive_herv$LZ),
                x = 'log2FoldChange',
                y = 'pvalue',
                title = 'NB vs LZ')
dev.off()

pdf("plots/05e-gcb_naive_volcano_NB_v_all.pdf", height=8, width=8)
EnhancedVolcano(sig_naive_herv$NB,
                lab = rownames(sig_naive_herv$NB),
                x = 'log2FoldChange',
                y = 'pvalue',
                title = 'All vs NB')
dev.off()

pdf("plots/05e-gcb_naive_volcano_MB_v_NB.pdf", height=8, width=8)
EnhancedVolcano(sig_naive_herv$MB,
                lab = rownames(sig_naive_herv$MB),
                x = 'log2FoldChange',
                y = 'pvalue',
                title = 'NB vs MB')
dev.off()

############################### FAMILY LEVEL UP ################################

upreg_hervs_nb_df <- do.call(rbind, lapply(upvars_naive_hervs, data.frame))
colnames(upreg_hervs_nb_df) <- c("herv")
upreg_hervs_nb_df$cell_type <- rownames(upreg_hervs_nb_df)
upreg_hervs_nb_df$cell_type <- gsub("\\..*","",upreg_hervs_nb_df$cell_type)
upreg_hervs_nb_df$family <- retro.annot$family[match(upreg_hervs_nb_df$herv, 
                                                         retro.annot$locus)]

upreg_families_naive <-
  upreg_hervs_nb_df %>% dplyr::count(family, cell_type, sort = TRUE) 

upreg_families_naive<-upreg_families_naive[!(upreg_families_naive$cell_type=="GCB"),]

pdf("plots/05e-gcb_naive_upreg_families.pdf", height=8, width=6)
ggplot(upreg_families_naive, aes(fill=reorder(family, -n), y=cell_type, x=n)) + 
  geom_bar(position="fill", stat="identity", colour="black", size=0.3) + 
  scale_fill_manual(values = c(pal_futurama("planetexpress")(12), 
                               pal_npg("nrc", alpha = 0.7)(10),
                               pal_jco("default", alpha=0.7)(10),
                               pal_nejm("default", alpha=0.7)(8),
                               pal_tron("legacy", alpha=0.7)(7),
                               pal_lancet("lanonc", alpha=0.7)(9),
                               pal_startrek("uniform", alpha=0.7)(7)),
                    breaks = unique(retro.annot$family),
                    labels = unique(retro.annot$family)) + 
  
  coord_flip() +
  theme_pubclean() +  
  guides(fill=guide_legend(title="TE Family")) +
  ylab("Upregulated HERVs by Family") +
  xlab("Proportion of HERVs") + 
  theme(legend.position = c("right")) + 
  guides(fill = guide_legend(title = "HERV family", ncol = 2))
dev.off()

############################### FAMILY LEVEL DOWN ##############################

down_hervs_nb_df <- do.call(rbind, lapply(downvars_naive_hervs, data.frame))
colnames(down_hervs_nb_df) <- c("herv")
down_hervs_nb_df$cell_type <- rownames(down_hervs_nb_df)
down_hervs_nb_df$cell_type <- gsub("\\..*","",down_hervs_nb_df$cell_type)
down_hervs_nb_df$family <- retro.annot$family[match(down_hervs_nb_df$herv, 
                                                        retro.annot$locus)]

downreg_families_naive <-
  down_hervs_nb_df %>% dplyr::count(family, cell_type, sort = TRUE) 

downreg_families_naive<-downreg_families_naive[!(downreg_families_naive$cell_type=="GCB"),]

pdf("plots/05e-gcb_naive_downreg_families.pdf", height=8, width=6)
ggplot(downreg_families_naive, aes(fill=reorder(family, -n), y=cell_type, x=n)) + 
  geom_bar(position="fill", stat="identity", colour="black", size=0.3) + 
  scale_fill_manual(values = c(pal_futurama("planetexpress")(12), 
                               pal_npg("nrc", alpha = 0.7)(10),
                               pal_jco("default", alpha=0.7)(10),
                               pal_nejm("default", alpha=0.7)(8),
                               pal_tron("legacy", alpha=0.7)(7),
                               pal_lancet("lanonc", alpha=0.7)(9),
                               pal_startrek("uniform", alpha=0.7)(7)),
                    breaks = unique(retro.annot$family),
                    labels = unique(retro.annot$family)) + 
  
  coord_flip() +
  theme_pubclean() +  
  guides(fill=guide_legend(title="TE Family")) +
  ylab("Downregulated HERVs by Family") +
  xlab("Proportion of HERVs") + 
  theme(legend.position = c("right")) + 
  guides(fill = guide_legend(title = "HERV family", ncol = 2))
dev.off()


############################# MERGED FAMILY UP/DOWN ############################


updown_family <- do.call(rbind, (list(Upregulated = upreg_families_naive, 
                                      Dowregulated = downreg_families_naive)))
updown_family$expression <- rownames(updown_family)
updown_family$expression <- gsub("\\..*","",updown_family$expression)
rownames(updown_family)<-NULL

pdf("plots/05e-gcb_naive_updown_families.pdf", height=8, width=10)
ggplot(updown_family, aes(fill=reorder(family, -n), y=cell_type, x=n)) + 
  geom_bar(position="fill", stat="identity", colour="black", size=0.3) + 
  scale_fill_manual(values = c(pal_futurama("planetexpress")(12), 
                               pal_npg("nrc", alpha = 0.7)(10),
                               pal_jco("default", alpha=0.7)(10),
                               pal_nejm("default", alpha=0.7)(8),
                               pal_tron("legacy", alpha=0.7)(7),
                               pal_lancet("lanonc", alpha=0.7)(9),
                               pal_startrek("uniform", alpha=0.7)(7)),
                    breaks = unique(retro.annot$family),
                    labels = unique(retro.annot$family)) + 
  
  coord_flip() +
  theme_pubclean() +  
  guides(fill=guide_legend(title="TE Family")) +
  ylab("Downregulated HERVs by Family") +
  xlab("Proportion of HERVs") + 
  theme(legend.position = c("right")) + 
  guides(fill = guide_legend(title = "HERV family", ncol = 2)) +
  facet_wrap(~ expression, ncol = 2)
dev.off()

#################################### TE TYPE ###################################

rownames(upreg_hervs_nb_df) <- NULL
upreg_hervs_nb_df$type <- retro.annot.v2$TE_type[
  match(upreg_hervs_nb_df$herv, retro.annot.v2$Locus)]

upreg_hervs_nb_df<-upreg_hervs_nb_df[!(upreg_hervs_nb_df$cell_type=="DZvLZ"),]

upreg_naive_te_type <-
  upreg_hervs_nb_df %>% dplyr::count(type, cell_type, sort = TRUE) 

pdf("plots/05e-gcb_naive_upreg_hervs_type.pdf", height=4, width=4)
ggplot(upreg_naive_te_type, aes(fill=reorder(type, -n), y=cell_type, x=n)) + 
  geom_bar(position="stack", stat="identity", colour="black", size=0.3) + 
  scale_fill_manual(values = c(pal_aaas("default", alpha=0.7)(3))) + 
  coord_flip() +
  theme_pubclean() +  
  guides(fill=guide_legend(title="HERV Type")) +
  ylab("Upregulated HERVs by Type") +
  xlab("Number of HERVs") + 
  theme(legend.position = c("right")) +
  theme(aspect.ratio = 1)
dev.off()

rownames(down_hervs_nb_df) <- NULL
down_hervs_nb_df$type <- retro.annot.v2$TE_type[
  match(down_hervs_nb_df$herv, retro.annot.v2$Locus)]

down_hervs_nb_df<-down_hervs_nb_df[!(down_hervs_nb_df$cell_type=="DZvLZ"),]

downreg_naive_te_type <-
  down_hervs_nb_df %>% dplyr::count(type, cell_type, sort = TRUE) 

pdf("plots/05e-gcb_naive_downreg_hervs_type.pdf", height=4, width=4)
ggplot(downreg_naive_te_type, aes(fill=reorder(type, -n), y=cell_type, x=n)) + 
  geom_bar(position="stack", stat="identity", colour="black", size=0.3) + 
  scale_fill_manual(values = c(pal_aaas("default", alpha=0.7)(3))) + 
  coord_flip() +
  theme_pubclean() +  
  guides(fill=guide_legend(title="HERV Type")) +
  ylab("Downregulated HERVs by Type") +
  xlab("Number of HERVs") + 
  theme(legend.position = c("right")) +
  theme(aspect.ratio = 1)
dev.off()


################################### SAVE DATA ##################################

upvars_gcb <- upvars
upvars_gcb_hervs <- upvars_hervs
downvars_gcb <- downvars
downvars_gcb_hervs <- downvars_hervs
sig_herv_gcb <- sig_herv
sig_gcb <- sig

save(upvars_gcb, upvars_gcb_hervs, downvars_gcb, downvars_gcb_hervs,
     sig_herv_gcb, sig_gcb, file = "r_outputs/05e-gcb_vars.Rdata")   


save(upvars_no_gcb, upvars_hervs_no_gcb, downvars_no_gcb, downvars_hervs_no_gcb,
     sig_herv_no_gcb, sig_no_gcb, file = "r_outputs/05e_gcb_vars_no_gcb.Rdata")

save(upvars_naive, upvars_naive_hervs, downvars_naive, downvars_naive_hervs,
     sig_naive, sig_naive_herv, file = "r_outputs/05_gcb_naive_vars.Rdata")

load("r_outputs/05e-gcb_vars.Rdata")
load("r_outputs/05e_gcb_vars_no_gcb.Rdata")
