==> 05q-dlbcl_subtypes_deseq.R <==
################################################################################
################################################################################
################################################################################
################################################################################
###################### BURKITT LYMPHOMA HERV CLUSTER DESEQ #####################

## Plan:
## 1.

#################################### SETUP #####################################

library(knitr)
library(tidyverse)
library(matrixStats)
library(data.table)
library(PCAtools)
library(DESeq2)
library(ggplot2)
library(ggsci)
library(ggpubr)
library(edgeR)
library(ashr)
library(cowplot)
library(wesanderson)
library(UpSetR)
library(ComplexUpset)
library(EnhancedVolcano)
library(pheatmap)
library(fgsea)
library(viridis)

################################### LOAD DATA ##################################

load("r_outputs/02-DLBCL_filt_counts.Rdata")
load("r_outputs/05o-DLBCL_pca_ccp_clusters_metadata.Rdata")
load("r_outputs/01-refs.Rdata")

################################################################################
################################################################################
#################################### CLUST 7 ###################################
################################################################################
################################################################################


############################# FUNCTION SCREE PLOT ##############################

pdf("plots/05o-dlbcl_subtypes_biplot_clust.k7.pdf", height=5, width=6)
biplot(DLBCL.herv.pca.obj, 
       lab = NULL,
       showLoadings = FALSE,
       pointSize = 2, 
       ellipse = FALSE,
       colby = "clust.retro.k7",
       shape = "project", 
       shapekey = c("NCICCR-DLBCL" = 15, "TCGA-DLBC" = 8),
       colkey = ggsci::pal_npg(palette = c("nrc"))(9),
       legendPosition = "right") +
  theme_cowplot() +
  theme(aspect.ratio = 1)

dev.off()


#################################### DESEQ2 ####################################

# p value threshold 
lfc.cutoff <- 1.5
pval=0.001 

# Sanity check
stopifnot(all(colnames(DLBCL.filt.comb) == rownames(DLBCL_metadata)))


# Run Deseq2
DLBCL.k7.dds <- DESeq2::DESeqDataSetFromMatrix(DLBCL.filt.comb, 
                                               DLBCL_metadata, 
                                            ~ clust.retro.k7 + 0)


DLBCL.k7.dds <- DESeq2::DESeq(DLBCL.k7.dds, parallel=T)
DLBCL.k7.tform <- DESeq2::varianceStabilizingTransformation(DLBCL.k7.dds, blind=FALSE)

################################# DESEQ2 HERVs ################################# 

stopifnot(all(colnames(DLBCL.filt.herv) == rownames(DLBCL_metadata)))

DLBCL.k7.herv.dds <- DESeq2::DESeqDataSetFromMatrix(DLBCL.filt.herv, 
                                                    DLBCL_metadata, 
                                                 ~ clust.retro.k7 + 0)


DLBCL.k7.herv.dds <- DESeq2::DESeq(DLBCL.k7.herv.dds, parallel=T)
DLBCL.k7.herv.tform <- DESeq2::varianceStabilizingTransformation(DLBCL.k7.herv.dds, 
                                                              blind=FALSE)


############################## TOP GENES AND HERVS ############################# 

res.k7 <- list(
  "HC1" = DESeq2::results(DLBCL.k7.dds, contrast=c(+1, -1/6, -1/6, -1/6, -1/6, -1/6, -1/6), alpha=pval),
  "HC2" = DESeq2::results(DLBCL.k7.dds, contrast=c(-1/6, +1, -1/6, -1/6, -1/6, -1/6, -1/6), alpha=pval),
  "HC3" = DESeq2::results(DLBCL.k7.dds, contrast=c(-1/6, -1/6, +1, -1/6, -1/6, -1/6, -1/6), alpha=pval),
  "HC4" = DESeq2::results(DLBCL.k7.dds, contrast=c(-1/6, -1/6, -1/6, +1, -1/6, -1/6, -1/6), alpha=pval),
  "HC5" = DESeq2::results(DLBCL.k7.dds, contrast=c(-1/6, -1/6, -1/6, -1/6, +1, -1/6, -1/6), alpha=pval),
  "HC6" = DESeq2::results(DLBCL.k7.dds, contrast=c(-1/6, -1/6, -1/6, -1/6, -1/6, +1, -1/6), alpha=pval),
  "HC7" = DESeq2::results(DLBCL.k7.dds, contrast=c(-1/6, -1/6, -1/6, -1/6, -1/6, -1/6, +1), alpha=pval)
)

res.k7 <- lapply(res.k7, function(r) {
  r$display <- gene_table[rownames(r),]$display
  r$class <- gene_table[rownames(r),]$gene_type
  r
})


sig.k7 <- lapply(res.k7, function(r) {
  s <- subset(r, padj < pval & abs(log2FoldChange) > lfc.cutoff)
  s[order(s$padj),]
})

for (n in names(sig.k7)) {
  cat("\n#--- Contrast", n, "---#\n")
  summary(sig.k7[[n]])
}

for (n in names(sig.k7)) {
  cat("\n#--- Contrast", n, "---#\n")
  print(head(sig.k7[[n]]))
}

################################ TOP HERVS ONLY ################################ 

res.k7.herv <- list(
  "HC1" = DESeq2::results(DLBCL.k7.herv.dds, contrast=c(+1, -1/6, -1/6, -1/6, -1/6, -1/6, -1/6), alpha=pval),
  "HC2" = DESeq2::results(DLBCL.k7.herv.dds, contrast=c(-1/6, +1, -1/6, -1/6, -1/6, -1/6, -1/6), alpha=pval),
  "HC3" = DESeq2::results(DLBCL.k7.herv.dds, contrast=c(-1/6, -1/6, +1, -1/6, -1/6, -1/6, -1/6), alpha=pval),
  "HC4" = DESeq2::results(DLBCL.k7.herv.dds, contrast=c(-1/6, -1/6, -1/6, +1, -1/6, -1/6, -1/6), alpha=pval),
  "HC5" = DESeq2::results(DLBCL.k7.herv.dds, contrast=c(-1/6, -1/6, -1/6, -1/6, +1, -1/6, -1/6), alpha=pval),
  "HC6" = DESeq2::results(DLBCL.k7.herv.dds, contrast=c(-1/6, -1/6, -1/6, -1/6, -1/6, +1, -1/6), alpha=pval),
  "HC7" = DESeq2::results(DLBCL.k7.herv.dds, contrast=c(-1/6, -1/6, -1/6, -1/6, -1/6, -1/6, +1), alpha=pval)
)

res.k7.herv <- lapply(res.k7.herv, function(r) {
  r$display <- gene_table[rownames(r),]$display
  r$class <- gene_table[rownames(r),]$gene_type
  r
})


sig.k7.herv <- lapply(res.k7.herv, function(r) {
  s <- subset(r, padj < pval & abs(log2FoldChange) > lfc.cutoff)
  s[order(s$padj),]
})

for (n in names(sig.k7.herv)) {
  cat("\n#--- Contrast", n, "---#\n")
  summary(sig.k7.herv[[n]])
}

for (n in names(sig.k7.herv)) {
  cat("\n#--- Contrast", n, "---#\n")
  print(head(sig.k7.herv[[n]]))
}


############################ UPSET GENES & HERVS ###############################

upvars.DLBCL.k7 <- lapply(sig.k7[1:7], function(r) rownames(subset(r, log2FoldChange>0)))
downvars.DLBCL.k7 <- lapply(sig.k7[1:7], function(r) rownames(subset(r, log2FoldChange<0)))

pdf("plots/05q-DLBCL_k7_upset_upvars.pdf", height=5, width=7)

ComplexUpset::upset(fromList(upvars.DLBCL.k7), 
                    intersect = c(names(upvars.DLBCL.k7)),
                    intersections = list( 
                      c("HC1"), 
                      c("HC2"), 
                      c("HC3"),
                      c("HC4"), 
                      c("HC5"),
                      c("HC6"),
                      c("HC7")), 
                    queries = list(
                      upset_query(set=c("HC1"), 
                                  color = "#E64B35FF", 
                                  fill = "#E64B35FF"),
                      upset_query(set=c("HC2"), 
                                  color = "#4DBBD5FF", 
                                  fill = "#4DBBD5FF"),
                      upset_query(set=c("HC3"), 
                                  color = "#00A087FF", 
                                  fill = "#00A087FF"),
                      upset_query(set=c("HC4"), 
                                  color = "#3C5488FF", 
                                  fill = "#3C5488FF"),
                      upset_query(set=c("HC5"), 
                                  color = "#F39B7FFF", 
                                  fill = "#F39B7FFF"),
                      upset_query(set=c("HC6"), 
                                  color = "#8491B4FF", 
                                  fill = "#8491B4FF"),
                      upset_query(set=c("HC7"), 
                                  color = "#91D1C2FF", 
                                  fill = "#91D1C2FF")
                    ))

dev.off()

pdf("plots/05q-DLBCL_k7_upset_downvars.pdf", height=5, width=7)
ComplexUpset::upset(fromList(downvars.DLBCL.k7), 
                    intersect = c(names(upvars.DLBCL.k7)),
                    intersections = list( 
                      c("HC1"), 
                      c("HC2"), 
                      c("HC3"),
                      c("HC4"), 
                      c("HC5"),
                      c("HC6"),
                      c("HC7")), 
                    queries = list(
                      upset_query(set=c("HC1"), 
                                  color = "#E64B35FF", 
                                  fill = "#E64B35FF"),
                      upset_query(set=c("HC2"), 
                                  color = "#4DBBD5FF", 
                                  fill = "#4DBBD5FF"),
                      upset_query(set=c("HC3"), 
                                  color = "#00A087FF", 
                                  fill = "#00A087FF"),
                      upset_query(set=c("HC4"), 
                                  color = "#3C5488FF", 
                                  fill = "#3C5488FF"),
                      upset_query(set=c("HC5"), 
                                  color = "#F39B7FFF", 
                                  fill = "#F39B7FFF"),
                      upset_query(set=c("HC6"), 
                                  color = "#8491B4FF", 
                                  fill = "#8491B4FF"),
                      upset_query(set=c("HC7"), 
                                  color = "#91D1C2FF", 
                                  fill = "#91D1C2FF")
                    ))
dev.off()

up.binmat.DLBCL.k7 <- fromList(upvars.DLBCL.k7)
rn <- do.call(c, upvars.DLBCL.k7)
rn <- rn[!duplicated(rn)]
rownames(up.binmat.DLBCL.k7) <- rn
rm(rn)

dn.binmat.DLBCL.k7 <- fromList(downvars.DLBCL.k7)
rn <- do.call(c, downvars.DLBCL.k7)
rn <- rn[!duplicated(rn)]
rownames(dn.binmat.DLBCL.k7) <- rn
rm(rn)

upvars.DLBCL.k7.sub <-
  lapply(sig.k7[1:7], function(r) rownames(subset(r, log2FoldChange>0))[1:100])
up.binmat.DLBCL.k7.sub <- fromList(upvars.DLBCL.k7.sub)
rn <- do.call(c, upvars.DLBCL.k7.sub)
rn <- rn[!duplicated(rn)]
rownames(up.binmat.DLBCL.k7.sub) <- rn
rm(rn)


upvars.DLBCL.k7.sub2 <-
  lapply(sig.k7[1:7], function(r) rownames(subset(r, log2FoldChange>0))[1:10])
up.binmat.DLBCL.k7.sub2 <- fromList(upvars.DLBCL.k7.sub2)
rn <- do.call(c, upvars.DLBCL.k7.sub2)
rn <- rn[!duplicated(rn)]
rownames(up.binmat.DLBCL.k7.sub2) <- rn
rm(rn)

############################## UPSET HERVS ONLY ################################

upvars.DLBCL.k7.herv <- lapply(sig.k7.herv[1:7], function(r) rownames(subset(r, log2FoldChange>0)))
downvars.DLBCL.k7.herv <- lapply(sig.k7.herv[1:7], function(r) rownames(subset(r, log2FoldChange<0)))

pdf("plots/05q-DLBCL_k7_upset_upvars_hervs.pdf", height=5, width=7)

ComplexUpset::upset(fromList(upvars.DLBCL.k7.herv), 
                    intersect = c(names(upvars.DLBCL.k7.herv)),
                    intersections = list( 
                      c("HC1"), 
                      c("HC2"), 
                      c("HC3"),
                      c("HC4"), 
                      c("HC5"),
                      c("HC6"),
                      c("HC7")), 
                    queries = list(
                      upset_query(set=c("HC1"), 
                                  color = "#E64B35FF", 
                                  fill = "#E64B35FF"),
                      upset_query(set=c("HC2"), 
                                  color = "#4DBBD5FF", 
                                  fill = "#4DBBD5FF"),
                      upset_query(set=c("HC3"), 
                                  color = "#00A087FF", 
                                  fill = "#00A087FF"),
                      upset_query(set=c("HC4"), 
                                  color = "#3C5488FF", 
                                  fill = "#3C5488FF"),
                      upset_query(set=c("HC5"), 
                                  color = "#F39B7FFF", 
                                  fill = "#F39B7FFF"),
                      upset_query(set=c("HC6"), 
                                  color = "#8491B4FF", 
                                  fill = "#8491B4FF"),
                      upset_query(set=c("HC7"), 
                                  color = "#91D1C2FF", 
                                  fill = "#91D1C2FF")
                    ))

dev.off()

pdf("plots/05q-DLBCL_k7_upset_downvars_hervs.pdf", height=5, width=7)
ComplexUpset::upset(fromList(downvars.DLBCL.k7.herv), 
                    intersect = c(names(upvars.DLBCL.k7)),
                    intersections = list( 
                      c("HC1"), 
                      c("HC2"), 
                      c("HC3"),
                      c("HC4"), 
                      c("HC5"),
                      c("HC6"),
                      c("HC7")), 
                    queries = list(
                      upset_query(set=c("HC1"), 
                                  color = "#E64B35FF", 
                                  fill = "#E64B35FF"),
                      upset_query(set=c("HC2"), 
                                  color = "#4DBBD5FF", 
                                  fill = "#4DBBD5FF"),
                      upset_query(set=c("HC3"), 
                                  color = "#00A087FF", 
                                  fill = "#00A087FF"),
                      upset_query(set=c("HC4"), 
                                  color = "#3C5488FF", 
                                  fill = "#3C5488FF"),
                      upset_query(set=c("HC5"), 
                                  color = "#F39B7FFF", 
                                  fill = "#F39B7FFF"),
                      upset_query(set=c("HC6"), 
                                  color = "#8491B4FF", 
                                  fill = "#8491B4FF"),
                      upset_query(set=c("HC7"), 
                                  color = "#91D1C2FF", 
                                  fill = "#91D1C2FF")
                    ))
dev.off()

up.binmat.DLBCL.k7.hervs <- fromList(upvars.DLBCL.k7.herv)
rn <- do.call(c, upvars.DLBCL.k7.herv)
rn <- rn[!duplicated(rn)]
rownames(up.binmat.DLBCL.k7.hervs) <- rn
rm(rn)

dn.binmat.DLBCL.k7.hervs <- fromList(downvars.DLBCL.k7.herv)
rn <- do.call(c, downvars.DLBCL.k7.herv)
rn <- rn[!duplicated(rn)]
rownames(dn.binmat.DLBCL.k7.hervs) <- rn
rm(rn)

upvars.DLBCL.k7.herv.sub <-
  lapply(sig.k7.herv[1:7], function(r) rownames(subset(r, log2FoldChange>0))[1:8])
up.binmat.DLBCL.k7.hervs.sub <- fromList(upvars.DLBCL.k7.herv.sub)
rn <- do.call(c, upvars.DLBCL.k7.herv.sub)
rn <- rn[!duplicated(rn)]
rownames(up.binmat.DLBCL.k7.hervs.sub) <- rn
rm(rn)


################################# HEATMAPS #####################################
################################### SETUP ######################################

# Colors for cells
cols <- rgb_gsea(palette = c("default"), n = 14, alpha = 0.7, reverse = FALSE)

# Annotation column
df <- as.data.frame(colData(DLBCL.k7.dds)[,c("COO_class", "LymphGen_call",
                                             "EcoTyper_call", "clust.retro.k7")])

# Create colors for each group
annoCol <- list(COO_class = c("GCB" = "royalblue", 
                                "ABC" = "red3", 
                                "Unclass" = "lightblue", 
                                "Missing" = "grey"),
                EcoTyper_call = c("S1" = "#bcc779",
                                     "S2" = "#60aaec",
                                     "S3" = "#Ecaf60",
                                     "S4" = "#B789EE",
                                     "S5" = "#E0AF71",
                                     "Unassigned" = "grey",
                                     "Missing" = "grey"),
                LymphGen_call = c("A53" = "#e78bf0",
                                     "BN2" = "#e28743",
                                     "BN2/A53" = "#AF6C3A",
                                     "BN2/EZB" = "#c79875",
                                     "BN2/MCD" = "#7a4c29",
                                     "BN2/ST2" = "#58361d",
                                     "EZB" = "#3d3aaf",
                                     "EZB/A53" = "#7775c7",
                                     "EZB/MCD" = "#8b89cf",
                                     "EZB/N1/ST2/A53" = "#b1b0df",
                                     "EZB/ST2/A53" = "#d8d8ef",
                                     "MCD" = "#4d8347",
                                     "MCD/A53" = "#82a87e",
                                     "MCD/ST2" = "#b8cdb5",
                                     "N1" = "#e84646",
                                     "N1/ST2" = "#f19090",
                                     "Other" = "#e9f190",
                                     "ST2" = "#38baaa",
                                     "NA" = "grey"),
                clust.retro.k7 = c("HC1" = "#E64B35FF",
                                    "HC2" = "#4DBBD5FF",
                                    "HC3" = "#00A087FF",
                                    "HC4" = "#3C5488FF",
                                    "HC5" = "#F39B7FFF",
                                    "HC6" = "#8491B4FF",
                                    "HC7" = "#91D1C2FF")
)

######################### UPREGULATED IN ALL GROUPS ############################
top.genes <- rownames(up.binmat.DLBCL.k7)
top.genes.sub <- rownames(up.binmat.DLBCL.k7.sub)
top.genes.sub.sub <- rownames(up.binmat.DLBCL.k7.sub2)
top.hervs <- rownames(up.binmat.DLBCL.k7.hervs)
top.hervs.sub <- rownames(up.binmat.DLBCL.k7.hervs.sub)

pdf("plots/05q-DLBCL_k7_geneshervs_upregulated_all.pdf", height=10, width=10)
pheatmap(assay(DLBCL.k7.dds)[,c(ccp.obj[[8]]$consensusTree$order)][top.genes,], 
         main="Upregulated Genes and HERVs, all clusters",
         cluster_rows=TRUE,
         show_rownames=FALSE,
         show_colnames = FALSE,
         color = cols,
         scale="row",
         breaks=seq(-3,3,length.out=14),
         labels_row = gene_table[top.genes,]$display,
         cluster_cols=FALSE, 
         treeheight_row=0,
         annotation_col=df,
         annotation_colors = annoCol)
dev.off()

pdf("plots/05q-DLBCL_k7_geneshervs_upregulated_sub.pdf", height=10, width=10)
pheatmap(assay(DLBCL.k7.dds)[,c(ccp.obj[[8]]$consensusTree$order)][top.genes.sub,], 
         main="Upregulated Genes and HERVs, all clusters",
         cluster_rows=TRUE,
         show_rownames=FALSE,
         show_colnames = FALSE,
         color = cols,
         scale="row",
         breaks=seq(-3,3,length.out=14),
         labels_row = gene_table[top.genes,]$display,
         cluster_cols=FALSE, 
         treeheight_row=0,
         annotation_col=df,
         annotation_colors = annoCol)
dev.off()

pdf("plots/05q-DLBCL_k7_geneshervs_upregulated_subsub.pdf", height=10, width=10)
pheatmap(assay(DLBCL.k7.dds)[,c(ccp.obj[[8]]$consensusTree$order)][top.genes.sub.sub,], 
         main="Upregulated Genes and HERVs, all clusters",
         cluster_rows=TRUE,
         show_rownames=TRUE,
         show_colnames = FALSE,
         color = cols,
         scale="row",
         breaks=seq(-3,3,length.out=14),
         labels_row = gene_table[top.genes,]$display,
         cluster_cols=FALSE, 
         treeheight_row=0,
         annotation_col=df,
         annotation_colors = annoCol)
dev.off()

pdf("plots/05q-DLBCL_k7_hervs_upregulated_all.pdf", height=10, width=10)
pheatmap(assay(DLBCL.k7.dds)[,c(ccp.obj[[8]]$consensusTree$order)][top.hervs,], 
         main="Upregulated HERVs, all clusters",
         cluster_rows=TRUE,
         show_rownames=FALSE,
         show_colnames = FALSE,
         color = cols,
         scale="row",
         breaks=seq(-3,3,length.out=14),
         labels_row = gene_table[top.hervs,]$display,
         cluster_cols=FALSE, 
         treeheight_row=0,
         annotation_col=df,
         annotation_colors = annoCol)
dev.off()

pdf("plots/05q-DLBCL_k7_hervs_upregulated_sub.pdf", height=10, width=10)
pheatmap(assay(DLBCL.k7.dds)[,c(ccp.obj[[8]]$consensusTree$order)][top.hervs.sub,], 
         main="Upregulated HERVs, all clusters",
         cluster_rows=TRUE,
         show_rownames=TRUE,
         show_colnames = FALSE,
         color = cols,
         scale="row",
         breaks=seq(-3,3,length.out=14),
         labels_row = gene_table[top.hervs,]$display,
         cluster_cols=FALSE, 
         treeheight_row=0,
         annotation_col=df,
         annotation_colors = annoCol)
dev.off()

########################### DE HERVs PER GC SITE ###############################

makeheatmap <- function(topgenes, ...) {
  args <- list(...)
  mat <- assay(DLBCL.k7.tform)[unique(topgenes), ]
  rowdist <- as.dist((1 - cor(t(mat), method='spearman'))/2)
  
  annotation_col <- df
  cols <- rgb_gsea(palette = c("default"), n = 14, alpha = 0.7, reverse = FALSE)
  
  if (topgenes[1] %in% retro.annot.v2$Locus) {
    annoRow <- as.data.frame(retro.annot.v2[,c("TE_type", "Locus")])
    annoRow <- annoRow[topgenes,]
    annoRow <- subset(annoRow, select = -c(2))
    annotation_row <- annoRow
  } else {
    annotation_row <- NULL
  }
  
  pheatmap(mat,
           color=cols,
           scale="row", breaks=seq(-3,3,length.out=14),
           clustering_distance_rows = rowdist,
           clustering_method="average",
           annotation_col = df,
           annotation_colors = annoCol,
           labels_row = gene_table[topgenes,]$display,
           show_colnames=F, 
           show_rownames = T,
           fontsize = 8, fontsize_row = 6, border_color = NA,
           legend=T,
           treeheight_col=10,
           treeheight_row=10, 
           cutree_cols=4,
           ...
  )
}

for(clust in names(sig.k7.herv)) {
  tg <- rownames(sig.k7.herv[[clust]][1:75,])
  p <- makeheatmap(tg, main=paste0('DE in cluster ', clust))
  pdf(paste0("plots/05p-all_lymphoma_subtype_top_de_hervs_", clust, ".pdf"), height=7, width=7)
  print(p)
  dev.off()
}

for(clust in names(sig.k7)) {
  tg <- rownames(sig.k7[[clust]][1:75,])
  p <- makeheatmap(tg, main=paste0('DE in cluster ', clust))
  pdf(paste0("plots/05p-all_lymphoma_subtype_top_de_geeneshervs_", clust, ".pdf"), height=7, width=7)
  print(p)
  dev.off()
}

################################### PATHWAYS ###################################

pathways.hallmark <- gmtPathways("gsea/h.all.v2023.1.Hs.symbols.gmt")
pathways.immune <- gmtPathways("gsea/c7.immunesigdb.v2023.1.Hs.symbols.gmt")
pathways.bp <- gmtPathways("gsea/c5.go.bp.v2023.1.Hs.symbols.gmt")
pathways.kegg <- gmtPathways("gsea/c2.cp.kegg.v2023.1.Hs.symbols.gmt.txt")
pathways.biocarta <- gmtPathways("gsea/c2.cp.biocarta.v2023.1.Hs.symbols.gmt.txt")


################################ FGSEA FUNCTION ################################

make.fsgsea <- function(pathway, fgsea.res, clust_name, pathway_name) {
  
  fgsea.res$SYMBOL <- gene_table[rownames(fgsea.res),]$display
  
  fgsea.res <- as.data.frame(fgsea.res) %>% 
    dplyr::select(SYMBOL, stat) %>% 
    na.omit() %>% 
    distinct() %>% 
    group_by(SYMBOL) %>% 
    summarize(stat=mean(stat))
  
  fgsea.ranks <- deframe(fgsea.res)
  
  fgsea.out <- fgsea(pathways=pathway, 
                     stats=fgsea.ranks, 
                     nPermSimple = 10000,
                     eps=0)
  
  return(fgsea.out)
  
  assign(paste0(clust_name, ".", pathway_name, ".fgsea.out"), fgsea.out, envir = .GlobalEnv )
  }

################################ HALLMARK FGSEA ################################

fsgsea.hallmarks.k7 <- list(
  "HC1/ABC-PB.hallmark" = make.fsgsea(pathways.hallmark, res.k7$HC1, "HC1/ABC-PB", "hallmark"),
  "HC2/ABC-MB.hallmark" = make.fsgsea(pathways.hallmark, res.k7$HC2, "HC2/ABC-MB", "hallmark"),
  "HC3/GCB-LZ.hallmark" = make.fsgsea(pathways.hallmark, res.k7$HC3, "HC3/GCB-LZ", "hallmark"),
  "HC4/GCB-Like.hallmark" = make.fsgsea(pathways.hallmark, res.k7$HC4, "HC4/GCB-Like", "hallmark"),
  "HC5/PB-Like.hallmark" = make.fsgsea(pathways.hallmark, res.k7$HC5, "HC5/PB-Like", "hallmark"),
  "HC6/GCB.hallmark" = make.fsgsea(pathways.hallmark, res.k7$HC6, "HC6/GCB", "hallmark"),
  "HC7/HERVH.hallmark" = make.fsgsea(pathways.hallmark, res.k7$HC7, "HC7/HERVH", "hallmark")
)

pdf("plots/05q-DLBCL_k7_all_c_hallmarks.pdf", height=10, width=7)
for(clust in names(fsgsea.hallmarks.k7)) {
  fgseaResTidy <- fsgsea.hallmarks.k7[[clust]] %>%
    as_tibble() %>%
    arrange(desc(NES))
  
  rownames(fgseaResTidy) <- fgseaResTidy$pathway
  
  p <- ggplot(fgseaResTidy, aes(reorder(pathway, NES), NES)) +
    geom_col(aes(fill=padj<0.05)) +
    coord_flip() +
    labs(x="Pathway", y="Normalized Enrichment Score",
         title=clust) + 
    theme_minimal()
  
  print(p)
}
dev.off()

# Get NES 
fsgsea.hallmarks.k7.summ <- as.data.frame(do.call(cbind, 
                                                  lapply(fsgsea.hallmarks.k7, 
                                           function(x) x[, c("NES")])))

# Recode NES summary dataframe
rownames(fsgsea.hallmarks.k7.summ) <- fsgsea.hallmarks.k7$C1.hallmark$pathway
rownames(fsgsea.hallmarks.k7.summ) <- gsub("HALLMARK_","",rownames(fsgsea.hallmarks.k7.summ))
colnames(fsgsea.hallmarks.k7.summ) <- gsub(".hallmark.NES","",colnames(fsgsea.hallmarks.k7.summ))
fsgsea.hallmarks.k7.summ <- 
  fsgsea.hallmarks.k7.summ[rowSums(is.na(fsgsea.hallmarks.k7.summ)) != ncol(fsgsea.hallmarks.k7.summ), ]


# Get longform df
fsgsea.hallmarks.k7.summary <- rbindlist(fsgsea.hallmarks.k7, idcol = "index")

# Recode df
fsgsea.hallmarks.k7.summary$pathway <- gsub("HALLMARK_","",fsgsea.hallmarks.k7.summary$pathway)
fsgsea.hallmarks.k7.summary$pathway <- gsub("_"," ",fsgsea.hallmarks.k7.summary$pathway)
fsgsea.hallmarks.k7.summary$index <- gsub(".hallmark","",fsgsea.hallmarks.k7.summary$index)


pdf("plots/05q-DLBCL_k7_all_hallmarks_heatmap.pdf", height=10, width=5)
pheatmap(fsgsea.hallmarks.k7.summ, 
         cluster_rows=FALSE,
         show_rownames=TRUE,
         show_colnames = TRUE,
         color = viridis_pal()(10),
         cluster_cols=TRUE, 
         treeheight_row=0)
dev.off()

fsgsea.hallmarks.k7.summary$index[fsgsea.hallmarks.k7.summary$index == "C1"]

pdf("plots/05q-DLBCL_k7_all_hallmarks_bubble.pdf", height=10, width=7.5)
ggplot(fsgsea.hallmarks.k7.summary, aes(x = index, 
                                        y = pathway, 
                                        size = -log(padj), 
                                        color = NES)) +
  geom_point() +
  scale_size(name = "-log (P value)", range = c(1, 10)) + 
  theme_cowplot() +
  theme(axis.text.x = element_text(angle = 35, hjust = 1))+
  scale_colour_gradientn(colors = viridis_pal()(10)) +
  xlab("DLBCL HERV Cluster") +
  ylab("Hallmark Pathway") 
dev.off()

################################## BP FGSEA ##################################

fsgsea.BP.k7 <- list(
  "HC1/ABC-PB" = make.fsgsea(pathways.bp, res.k7$HC1, "HC1/ABC-PB", "bp"),
  "HC2/ABC-MB" = make.fsgsea(pathways.bp, res.k7$HC2, "HC2/ABC-MB", "bp"),
  "HC3/GCB-LZ" = make.fsgsea(pathways.bp, res.k7$HC3, "HC3/GCB-LZ", "bp"),
  "HC4/GCB-Like" = make.fsgsea(pathways.bp, res.k7$HC4, "HC4/GCB-Like", "bp"),
  "HC5/PB-Like" = make.fsgsea(pathways.bp, res.k7$HC5, "HC5/PB-Like", "bp"),
  "HC6/GCB" = make.fsgsea(pathways.bp, res.k7$HC6, "HC6/GCB", "bp"),
  "HC7/HERVH" = make.fsgsea(pathways.bp, res.k7$HC7, "HC7/HERVH", "bp")
)

# Get NES 
fsgsea.BP.k7.summ <- as.data.frame(do.call(cbind, 
                                           lapply(fsgsea.BP.k7, 
                                                  function(x) x[, c("NES")])))

# Recode NES summary dataframe
rownames(fsgsea.BP.k7.summ) <- fsgsea.BP.k7$C1$pathway
fsgsea.BP.k7.summ <- 
  fsgsea.BP.k7.summ[rowSums(is.na(fsgsea.BP.k7.summ)) != ncol(fsgsea.BP.k7.summ), ]


# Get longform df
fsgsea.BP.k7.summary <- rbindlist(fsgsea.BP.k7, idcol = "index")

pdf("plots/05q-DLBCL_k7_all_bp_bubble.pdf", height=50, width=7.5)
ggplot(fsgsea.BP.k7.summary, aes(x = index, 
                                        y = pathway, 
                                        size = -log(padj), 
                                        color = NES)) +
  geom_point() +
  scale_size(name = "-log (P value)", range = c(1, 10)) + 
  theme_cowplot() +
  theme(axis.text.x = element_text(angle = 35, hjust = 1))+
  scale_colour_gradientn(colors = viridis_pal()(10)) +
  xlab("DLBCL HERV Cluster") +
  ylab("GO BP Pathway") 
dev.off()

# top 10 pathways per cluster

top10.bp <- fsgsea.BP.k7.summary %>% 
  arrange(padj) %>% 
  group_by(index) %>% dplyr::slice(1:10)

top10.bp.pathways <- unique(top10.bp$pathway)
top10.bp.pathways <- subset(fsgsea.BP.k7.summary, pathway %in% top10.bp.pathways)
top10.bp.pathways$pathway <- gsub("GOBP_","",top10.bp.pathways$pathway)
top10.bp.pathways$pathway <- gsub("_"," ",top10.bp.pathways$pathway)

pdf("plots/05q-DLBCL_k7_all_topbp_bubble.pdf", height=13, width=11)
ggplot(top10.bp.pathways, aes(x = index, 
                                 y = pathway, 
                                 size = -log(padj), 
                                 color = NES)) +
  geom_point() +
  scale_size(name = "-log (P value)", range = c(1, 10)) + 
  theme_cowplot() +
  theme(axis.text.x = element_text(angle = 35, hjust = 1))+
  scale_colour_gradientn(colors = viridis_pal()(10)) +
  xlab("DLBCL HERV Cluster") +
  ylab("GO BP Pathway") 
dev.off()

################################# IMMUNE FGSEA #################################

fsgsea.immune.k7 <- list(
  "C1" = make.fsgsea(pathways.immune, res.k7$C1, "C1", "immune"),
  "C2" = make.fsgsea(pathways.immune, res.k7$C2, "C2", "immune"),
  "C3" = make.fsgsea(pathways.immune, res.k7$C3, "C3", "immune"),
  "C4" = make.fsgsea(pathways.immune, res.k7$C4, "C4", "immune"),
  "C5" = make.fsgsea(pathways.immune, res.k7$C5, "C5", "immune"),
  "C6" = make.fsgsea(pathways.immune, res.k7$C6, "C6", "immune"),
  "C7" = make.fsgsea(pathways.immune, res.k7$C7, "C7", "immune")
)

pdf("plots/05q-DLBCL_k7_all_c_immune.pdf", height=10, width=7)
for(clust in names(fsgsea.immune.k7)) {
  fgseaResTidy.immune <- fsgsea.immune.k7[[clust]] %>%
    as_tibble() %>%
    arrange(desc(NES))
  
  rownames(fgseaResTidy.immune) <- fgseaResTidy.immune$pathway
  fgseaResTidy.immune <- fgseaResTidy.immune[fgseaResTidy.immune$pathway %like% "GC", ]
  
  p <- ggplot(fgseaResTidy.immune, aes(reorder(pathway, NES), NES)) +
    geom_col(aes(fill=padj<0.05)) +
    coord_flip() +
    labs(x="Pathway", y="Normalized Enrichment Score",
         title=clust) + 
    theme_minimal()
  
  print(p)
}
dev.off()

# Get longform df
fsgsea.immune.k7.summary <- rbindlist(fsgsea.immune.k7, idcol = "index")

# top 10 pathways per cluster

top10.immune <- fsgsea.immune.k7.summary %>% 
  arrange(padj) %>% 
  group_by(index) %>% dplyr::slice(1:10)

top10.immune.pathways <- unique(top10.immune$pathway)
top10.immune.pathways <- subset(fsgsea.immune.k7.summary, pathway %in% top10.immune.pathways)
top10.immune.pathways$pathway <- gsub("_"," ",top10.immune.pathways$pathway)

pdf("plots/05q-DLBCL_k7_all_topimmune_bubble.pdf", height=13, width=13)
ggplot(top10.immune.pathways, aes(x = index, 
                              y = pathway, 
                              size = -log(padj), 
                              color = NES)) +
  geom_point() +
  scale_size(name = "-log (P value)", range = c(1, 10)) + 
  theme_cowplot() +
  theme(axis.text.x = element_text(angle = 35, hjust = 1))+
  scale_colour_gradientn(colors = viridis_pal()(10)) +
  xlab("DLBCL HERV Cluster") +
  ylab("Immune Pathway") 
dev.off()
################################## KEGG FGSEA ##################################

fsgsea.kegg.k7 <- list(
  "C1" = make.fsgsea(pathways.kegg, res.k7$C1, "C1", "KEGG"),
  "C2" = make.fsgsea(pathways.kegg, res.k7$C2, "C2", "KEGG"),
  "C3" = make.fsgsea(pathways.kegg, res.k7$C3, "C3", "KEGG"),
  "C4" = make.fsgsea(pathways.kegg, res.k7$C4, "C4", "KEGG"),
  "C5" = make.fsgsea(pathways.kegg, res.k7$C5, "C5", "KEGG"),
  "C6" = make.fsgsea(pathways.kegg, res.k7$C6, "C6", "KEGG"),
  "C7" = make.fsgsea(pathways.kegg, res.k7$C7, "C7", "KEGG")
)

pdf("plots/05q-DLBCL_k7_all_c_kegg.pdf", height=20, width=7)
for(clust in names(fsgsea.kegg.k7)) {
  fgseaResTidy.kegg <- fsgsea.kegg.k7[[clust]] %>%
    as_tibble() %>%
    arrange(desc(NES))
  
  rownames(fsgsea.kegg.k7[[clust]]) <- fsgsea.kegg.k7[[clust]]$pathway
  
  rownames(fgseaResTidy.kegg) <- fgseaResTidy.kegg$pathway
  
  p <- ggplot(fgseaResTidy.kegg, aes(reorder(pathway, NES), NES)) +
    geom_col(aes(fill=padj<0.05)) +
    coord_flip() +
    labs(x="Pathway", y="Normalized Enrichment Score",
         title=clust) + 
    theme_minimal()
  
  print(p)
  
  }
dev.off()

# Get longform df
fsgsea.kegg.k7.summary <- rbindlist(fsgsea.kegg.k7, idcol = "index")

# top 10 pathways per cluster

top10.kegg <- fsgsea.kegg.k7.summary %>% 
  arrange(padj) %>% 
  group_by(index) %>% dplyr::slice(1:10)

top10.kegg.pathways <- unique(top10.kegg$pathway)
top10.kegg.pathways <- subset(fsgsea.kegg.k7.summary, pathway %in% top10.kegg.pathways)
top10.kegg.pathways$pathway <- gsub("_"," ",top10.kegg.pathways$pathway)
top10.kegg.pathways$pathway <- gsub("KEGG "," ",top10.kegg.pathways$pathway)

pdf("plots/05q-DLBCL_k7_all_topkegg_bubble.pdf", height=10, width=8)
ggplot(top10.kegg.pathways, aes(x = index, 
                                  y = pathway, 
                                  size = -log(padj), 
                                  color = NES)) +
  geom_point() +
  scale_size(name = "-log (P value)", range = c(1, 10)) + 
  theme_cowplot() +
  theme(axis.text.x = element_text(angle = 35, hjust = 1))+
  scale_colour_gradientn(colors = viridis_pal()(10)) +
  xlab("DLBCL HERV Cluster") +
  ylab("KEGG Pathway") 
dev.off()

################################ BIOCARTA FGSEA ################################

fsgsea.biocarta.k7 <- list(
  "HC1/ABC-PB" = make.fsgsea(pathways.biocarta, res.k7$HC1, "HC1/ABC-PB", "Biocarta"),
  "HC2/ABC-MB" = make.fsgsea(pathways.biocarta, res.k7$HC2, "HC2/ABC-MB", "Biocarta"),
  "HC3/GCB-LZ" = make.fsgsea(pathways.biocarta, res.k7$HC3, "HC3/GCB-LZ", "Biocarta"),
  "HC4/GCB-Like" = make.fsgsea(pathways.biocarta, res.k7$HC4, "HC4/GCB-Like", "Biocarta"),
  "HC5/PB-Like" = make.fsgsea(pathways.biocarta, res.k7$HC5, "HC5/PB-Like", "Biocarta"),
  "HC6/GCB" = make.fsgsea(pathways.biocarta, res.k7$HC6, "HC6/GCB", "Biocarta"),
  "HC7/HERVH" = make.fsgsea(pathways.biocarta, res.k7$HC7, "HC7/HERVH", "Biocarta")
)


pdf("plots/05q-DLBCL_k7_all_c_biocarta.pdf", height=30, width=7)
for(clust in names(fsgsea.biocarta.k7)) {
  fgseaResTidy.biocarta <- fsgsea.biocarta.k7[[clust]] %>%
    as_tibble() %>%
    arrange(desc(NES))
  
  rownames(fsgsea.biocarta.k7[[clust]]) <- fsgsea.biocarta.k7[[clust]]$pathway
  
  rownames(fgseaResTidy.biocarta) <- fgseaResTidy.biocarta$pathway
  
  p <- ggplot(fgseaResTidy.biocarta, aes(reorder(pathway, NES), NES)) +
    geom_col(aes(fill=padj<0.05)) +
    coord_flip() +
    labs(x="Pathway", y="Normalized Enrichment Score",
         title=clust) + 
    theme_minimal()
  
  print(p)
  
}
dev.off()

for(clust in names(fsgsea.biocarta.k7)) {
  
  fsgsea.biocarta.k7[[clust]]$NES[fsgsea.biocarta.k7[[clust]]$padj >= 0.05] <- NA 
}

fgseaResTidy.biocarta.summ <- as.data.frame(do.call(cbind, 
                                                lapply(fsgsea.biocarta.k7, 
                                                       function(x) x[, c("NES")])))
rownames(fgseaResTidy.biocarta.summ) <- fsgsea.biocarta.k7$C1$pathway

colnames(fgseaResTidy.biocarta.summ) <- gsub(".NES","",colnames(fgseaResTidy.biocarta.summ))
fgseaResTidy.biocarta.summ <- 
  fgseaResTidy.biocarta.summ[rowSums(is.na(fgseaResTidy.biocarta.summ)) != ncol(fgseaResTidy.biocarta.summ), ]

pdf("plots/05q-DLBCL_k7_biocarta_heatmap.pdf", height=40, width=10)
pheatmap(data.matrix(fgseaResTidy.biocarta.summ), 
         cluster_rows=FALSE,
         show_rownames=TRUE,
         show_colnames = TRUE,
         color = viridis_pal()(10),
         cluster_cols=FALSE, 
         treeheight_row=0)
dev.off()


# Get longform df
fsgsea.biocarta.k7.summary <- rbindlist(fsgsea.biocarta.k7, idcol = "index")

# top 10 pathways per cluster

top10.biocarta <- fsgsea.biocarta.k7.summary %>% 
  arrange(padj) %>% 
  group_by(index) %>% dplyr::slice(1:10)

top10.biocarta.pathways <- unique(top10.biocarta$pathway)
top10.biocarta.pathways <- subset(fsgsea.biocarta.k7.summary, pathway %in% top10.biocarta.pathways)
top10.biocarta.pathways$pathway <- gsub("BIOCARTA_","",top10.biocarta.pathways$pathway)
top10.biocarta.pathways$pathway <- gsub("_"," ",top10.biocarta.pathways$pathway)

pdf("plots/05q-DLBCL_k7_all_topbiocarta_bubble.pdf", height=11, width=6.5)
ggplot(top10.biocarta.pathways, aes(x = index, 
                              y = pathway, 
                              size = -log(padj), 
                              color = NES)) +
  geom_point() +
  scale_size(name = "-log (P value)", range = c(1, 10)) + 
  theme_cowplot() +
  theme(axis.text.x = element_text(angle = 35, hjust = 1))+
  scale_colour_gradientn(colors = viridis_pal()(10)) +
  xlab("DLBCL HERV Cluster") +
  ylab("BioCarta Pathway") 
dev.off()

################################# B CELL FGSEA #################################

load("r_outputs/05f-agirre_vars.Rdata")

gene.sets.b.cell <- lapply(upvars_agirre[1:6], function(r) gene_table[r[1:150],]$display)
herv.sets.b.cell <- lapply(upvars_agirre_hervs[1:6], function(r) r[1:25])

gene.herv.sets.b.cell <- list("DZ" = append(gene.sets.b.cell$DZ, herv.sets.b.cell$DZ),
                              "LZ" = append(gene.sets.b.cell$LZ, herv.sets.b.cell$LZ),
                              "PB" = append(gene.sets.b.cell$PB, herv.sets.b.cell$PB),
                              "BMPC" = append(gene.sets.b.cell$BMPC, herv.sets.b.cell$BMPC),
                              "NB" = append(gene.sets.b.cell$NB, herv.sets.b.cell$NB),
                              "MB" = append(gene.sets.b.cell$MB, herv.sets.b.cell$MB)
                              )
 
write.table(as.data.frame(do.call(cbind, gene.herv.sets.b.cell)), 
            file = "r_outputs/gene.herv.sets.b.cell.txt",
            sep = "\t",quote = FALSE,
            col.names = TRUE,
            row.names = FALSE)

b.cell.k7 <- list(
  "HC1/ABC-PB" = make.fsgsea(gene.herv.sets.b.cell, res.k7$HC1, "HC1/ABC-PB", "b.cell"),
  "HC2/ABC-MB" = make.fsgsea(gene.herv.sets.b.cell, res.k7$HC2, "HC2/ABC-MB", "b.cell"),
  "HC3/GCB-LZ" = make.fsgsea(gene.herv.sets.b.cell, res.k7$HC3, "HC3/GCB-LZ", "b.cell"),
  "HC4/GCB-Like" = make.fsgsea(gene.herv.sets.b.cell, res.k7$HC4, "HC4/GCB-Like", "b.cell"),
  "HC5/PB-Like" = make.fsgsea(gene.herv.sets.b.cell, res.k7$HC5, "HC5/PB-Like", "b.cell"),
  "HC6/GCB" = make.fsgsea(gene.herv.sets.b.cell, res.k7$HC6, "HC6/GCB", "b.cell"),
  "HC7/HERVH" = make.fsgsea(gene.herv.sets.b.cell, res.k7$HC7, "HC7/HERVH", "b.cell")
)

pdf("plots/05q-DLBCL_k7_all_c_bcell.pdf", height=10, width=7)
for(clust in names(b.cell.k7)) {
  fgseaResTidy <- b.cell.k7[[clust]] %>%
    as_tibble() %>%
    arrange(desc(NES))
  
  rownames(fgseaResTidy) <- fgseaResTidy$pathway
  
  p <- ggplot(fgseaResTidy, aes(reorder(pathway, NES), NES)) +
    geom_col(aes(fill=padj<0.05)) +
    coord_flip() +
    labs(x="Pathway", y="Normalized Enrichment Score",
         title=clust) + 
    theme_minimal()
  
  print(p)
}
dev.off()

# Get longform df
fsgsea.b.cell.k7.summary <- rbindlist(b.cell.k7, idcol = "index")


pdf("plots/05q-DLBCL_k7_bcell_bubble.pdf", height=5.2, width=6)
ggplot(fsgsea.b.cell.k7.summary, aes(x = index, 
                                        y = pathway, 
                                        size = -log(padj), 
                                        color = NES)) +
  geom_point() +
  scale_size(name = "-log (P value)", range = c(1, 15)) + 
  theme_cowplot() +
  theme(axis.text.x = element_text(angle = 35, hjust = 1))+
  scale_colour_gradientn(colors = viridis_pal()(10)) +
  xlab("DLBCL HERV Cluster") +
  ylab("B Cell Signature") 
dev.off()

################################# PLOT COUNTS ################################## 


plot.counts <- function(df, gene, title) {
  
  as.data.frame(plotCounts(df, 
                           gene=gene, 
                           intgroup="clust.retro.k7", 
                           returnData = TRUE)) %>%
    ggplot(aes(x=clust.retro.k7, y=count, fill=clust.retro.k7))  +
    geom_boxplot(notch = TRUE) +
    theme_pubr() +
    theme(legend.position="none",
          axis.text.x = element_text(angle=45, hjust=1)) +
    xlab("HERV Cluster") +
    ylab("Counts") +
    ggtitle(title) + 
    theme(plot.title = element_text(hjust = 0.5),
          aspect.ratio = 1) +
    scale_fill_manual(values = c("C1" = "#E64B35FF",
                                 "C2" = "#4DBBD5FF",
                                 "C3" = "#00A087FF",
                                 "C4" = "#3C5488FF",
                                 "C5" = "#F39B7FFF",
                                 "C6" = "#8491B4FF",
                                 "C7" = "#91D1C2FF")) + 
    scale_y_log10(labels = label_comma()) 
}

plot.counts(DLBCL.k7.dds, "ENSG00000143379.12", "SETDB1") 


plot.counts(DLBCL.k7.dds, "HARLEQUIN_1q32.1", "HARLEQUIN_1q32.1") 
plot.counts(DLBCL.k7.dds, "HARLEQUIN_17q25.3b", "HARLEQUIN_17q25.3b") 
plot.counts(DLBCL.k7.dds, "HML2_1q22", "HML2_1q22")
plot.counts(DLBCL.k7.dds, "HML5_1q22", "HML5_1q22")

# C2
plot.counts(DLBCL.k7.dds, "HARLEQUIN_4p15.2", "HARLEQUIN_4p15.2")
plot.counts(DLBCL.k7.dds, "ENSG00000204531.20", "POU5F1")
plot.counts(DLBCL.k7.dds, "ENSG00000101336.18", "HCK")
plot.counts(DLBCL.k7.dds, "ENSG00000148843.15", "CD11C")


plot.counts(DLBCL.k7.dds, "ENSG00000181449.4", "SOX2")
plot.counts(DLBCL.k7.dds, "ENSG00000136826.15", "KLF4")

plot.counts(DLBCL.k7.dds, "ENSG00000182578.14", "CSF1R") 
plot.counts(DLBCL.k7.dds, "ENSG00000187621.15", "TCL6")
plot.counts(DLBCL.k7.dds, "ENSG00000214595.12", "EML6")

pdf("plots/05q-dlbcl_subtypes_fabp7.pdf", height=4,width=4)
plot.counts(DLBCL.k7.dds, "ENSG00000164434.12", "FABP7") +
  stat_compare_means(comparisons = list(c("C1", "C2"),
                                        c("C2", "C3"),
                                        c("C2", "C4"),
                                        c("C2", "C5"),
                                        c("C2", "C7"),
                                        c("C2", "C6")),
                     method = "t.test", 
                     label = "p.signif")
dev.off()

pdf("plots/05q-dlbcl_subtypes_prdm15.pdf", height=4,width=4)
plot.counts(DLBCL.k7.dds, "ENSG00000141956.14", "PRDM15") +
  stat_compare_means(comparisons = list(c("C5", "C1"),
                                        c("C5", "C2"),
                                        c("C5", "C3"),
                                        c("C5", "C6"),
                                        c("C5", "C7")),
                     method = "t.test", 
                     label = "p.signif")
dev.off()

plot.counts(DLBCL.k7.dds, "ENSG00000057657.17", "PRDM1") +
  stat_compare_means(comparisons = list(c("C5", "C1"),
                                        c("C5", "C2"),
                                        c("C5", "C3"),
                                        c("C5", "C6"),
                                        c("C5", "C7")),
                     method = "t.test", 
                     label = "p.signif")
plot.counts(DLBCL.k7.dds, "ENSG00000116731.22", "PRDM2")+
  stat_compare_means(comparisons = list(c("C1", "C2"),
                                        c("C2", "C4"),
                                        c("C4", "C5")),
                     method = "t.test", 
                     label = "p.signif")

plot.counts(DLBCL.k7.dds, "ENSG00000152784.16", "PRDM8") +
  stat_compare_means(comparisons = list(c("C1", "C2"),
                                        c("C4", "C5"),
                                        c("C2", "C5"),
                                        c("C5", "C7")),
                     method = "t.test", 
                     label = "p.signif")



plot.counts(DLBCL.k7.dds, "ENSG00000171791.14", "BCL2")
plot.counts(DLBCL.k7.dds, "ENSG00000171791.14", "BLIMP1")

################################# FAMILY LEVEL ################################# 

upreg.hervs.df <- do.call(rbind, lapply(upvars.DLBCL.k7.herv, data.frame))
colnames(upreg.hervs.df) <- c("herv")
upreg.hervs.df$cancer_type <- rownames(upreg.hervs.df)
upreg.hervs.df$cancer_type <- gsub("\\..*","",upreg.hervs.df$cancer_type)
upreg.hervs.df$family <- retro.annot$family[match(upreg.hervs.df$herv, 
                                                  retro.annot$locus)]

upreg.families <-
  upreg.hervs.df %>% dplyr::count(family, cancer_type, sort = TRUE) 

down.hervs.df <- do.call(rbind, lapply(downvars.DLBCL.k7.herv, data.frame))
colnames(down.hervs.df) <- c("herv")
down.hervs.df$cancer_type <- rownames(down.hervs.df)
down.hervs.df$cancer_type <- gsub("\\..*","",down.hervs.df$cancer_type)
down.hervs.df$family <- retro.annot$family[match(down.hervs.df$herv, 
                                                 retro.annot$locus)]

downreg.families <-
  down.hervs.df %>% dplyr::count(family, cancer_type, sort = TRUE) 

updown.family <- do.call(rbind, (list(Upregulated = upreg.families, 
                                      Dowregulated = downreg.families)))
updown.family$expression <- rownames(updown.family)
updown.family$expression <- gsub("\\..*","",updown.family$expression)
rownames(updown.family)<-NULL

pdf("plots/05q-dlbcl_updown_families_all.pdf", height=8, width=8)
ggplot(updown.family, aes(fill=reorder(family, -n), y=cancer_type, x=n)) + 
  geom_bar(position="fill", stat="identity", colour="black", size=0.3) + 
  scale_fill_manual(values = c(pal_futurama("planetexpress")(12), 
                               pal_npg("nrc", alpha = 0.7)(10),
                               pal_jco("default", alpha=0.7)(10),
                               pal_nejm("default", alpha=0.7)(8),
                               pal_tron("legacy", alpha=0.7)(7),
                               pal_lancet("lanonc", alpha=0.7)(9),
                               pal_startrek("uniform", alpha=0.7)(7)),
                    breaks = unique(retro.annot$family),
                    labels = unique(retro.annot$family)) + 
  coord_flip() +
  theme_cowplot() +  
  theme(axis.text.x = element_text(angle=45, hjust=1)) +
  guides(fill=guide_legend(title="TE Family")) +
  ylab(NULL) +
  xlab("Number of HERV Loci") + 
  theme(legend.position = c("right"),
        plot.margin = margin(10, 10, 10, 40),
        axis.line=element_blank()) + 
  guides(fill = guide_legend(title = "HERV family", ncol = 2)) +
  facet_wrap(~ expression, ncol = 2)
dev.off()
